% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting.R, R/qc_and_rollup.R
\name{apafTheme}
\alias{apafTheme}
\title{ProCan ggplot2 theme. Rectangle box around each plot.}
\usage{
apafTheme()

apafTheme()
}
\description{
Standard ggplot2 theme for ProCan plots

Standard ggplot2 theme for ProCan plots
}
