% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny_logging.R
\name{appender_shiny}
\alias{appender_shiny}
\title{Appender function for Shiny}
\usage{
appender_shiny(lines)
}
\arguments{
\item{lines}{A character vector of the log message.}
}
\description{
This function is a custom log appender for the 'logger' package.
It captures log records, formats them, and appends them to a reactiveVal
for display in a Shiny UI.
}
