% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_replicates.R
\name{averageTechReps,ProteinQuantitativeData-method}
\alias{averageTechReps,ProteinQuantitativeData-method}
\title{Average Technical Replicates}
\usage{
\S4method{averageTechReps}{ProteinQuantitativeData}(theObject, design_matrix_columns = c())
}
\arguments{
\item{theObject}{A ProteinQuantitativeData object}

\item{design_matrix_columns}{Additional columns to keep in the design matrix}
}
\value{
The modified ProteinQuantitativeData object with averaged technical replicates
}
\description{
Averages protein intensities across technical replicates, collapsing the
design matrix accordingly.
}
