% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_rollup_count_helpers.R
\name{calcPeptidesPerProtein}
\alias{calcPeptidesPerProtein}
\title{Calculate peptides per protein}
\usage{
calcPeptidesPerProtein(data)
}
\arguments{
\item{data}{A data frame or S4 object containing protein/peptide data}
}
\value{
Data frame with protein IDs and peptide counts
}
\description{
Calculate peptides per protein
}
