% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_data_helpers.R
\name{changeToCategorical}
\alias{changeToCategorical}
\title{getOneContinousPalette}
\usage{
changeToCategorical(metadata_tbl, column_name, num_colours = 9)
}
\arguments{
\item{metadata_tbl, column_name, num_colours}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
getOneContinousPalette
}
