% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_accession_methods.R
\name{chooseBestProteinAccession,ProteinQuantitativeData-method}
\alias{chooseBestProteinAccession,ProteinQuantitativeData-method}
\title{Choose Best Protein Accession for ProteinQuantitativeData}
\usage{
\S4method{chooseBestProteinAccession}{ProteinQuantitativeData}(
  theObject,
  delim = NULL,
  seqinr_obj = NULL,
  seqinr_accession_column = NULL,
  replace_zero_with_na = NULL,
  aggregation_method = NULL
)
}
\arguments{
\item{theObject}{The object of class ProteinQuantitativeData}

\item{delim}{The delimiter used to split the protein accessions}

\item{seqinr_obj}{The object of class Seqinr::seqinr}

\item{seqinr_accession_column}{The column in the seqinr object that contains the protein accessions}

\item{replace_zero_with_na}{Replace zero values with NA}

\item{aggregation_method}{Method to aggregate protein values: "sum", "mean", or "median" (default: "sum")}
}
\description{
Choose Best Protein Accession for ProteinQuantitativeData
}
