% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_design_control_methods.R
\name{cleanDesignMatrix,MetaboliteAssayData-method}
\alias{cleanDesignMatrix,MetaboliteAssayData-method}
\title{Clean Design Matrix for MetaboliteAssayData}
\usage{
\S4method{cleanDesignMatrix}{MetaboliteAssayData}(theObject)
}
\arguments{
\item{theObject}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Clean Design Matrix for MetaboliteAssayData
}
