% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{countStatDeGenes}
\alias{countStatDeGenes}
\title{Count the number of statistically significant differentially expressed proteins (according to user-defined threshold)}
\usage{
countStatDeGenes(
  data,
  lfc_thresh = 0,
  q_val_thresh = 0.05,
  log_fc_column = log2FC,
  q_value_column = fdr_qvalue
)
}
\arguments{
\item{lfc_thresh}{A numerical value specifying the log fold-change threhold (absolute value) for calling statistically significant proteins.}

\item{q_val_thresh}{A numerical value specifying the q-value threshold for statistically significant proteins.}

\item{log_fc_column}{The name of the log fold-change column (tidyverse style).}

\item{q_value_column}{The name of the q-value column (tidyverse style).}
}
\value{
A table with the following columns:
status: The status could be Significant and Up, Significant and Down or Not significant
counts: The number of proteins wit this status
}
\description{
Count the number of statistically significant differentially expressed proteins (according to user-defined threshold)
}
