% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_s4_objects.R
\name{createDAResultsForEnrichment}
\alias{createDAResultsForEnrichment}
\title{Create DA Results For Enrichment}
\usage{
createDAResultsForEnrichment(contrasts_tbl, design_matrix, da_output_dir)
}
\arguments{
\item{contrasts_tbl}{A tibble containing contrast information}

\item{design_matrix}{A data frame containing the design matrix}

\item{da_output_dir}{Directory containing DA results files}
}
\value{
An S4 object of class da_results_for_enrichment
}
\description{
Create DA Results For Enrichment
}
