% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_s4_objects.R
\name{createEnrichmentResults}
\alias{createEnrichmentResults}
\title{Create EnrichmentResults Object}
\usage{
createEnrichmentResults(contrasts_tbl)
}
\arguments{
\item{contrasts_tbl}{A tibble containing contrast information}
}
\value{
An S4 object of class EnrichmentResults
}
\description{
Create EnrichmentResults Object
}
