% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_da_handlers_load_session.R
\name{da_server_load_session_handler}
\alias{da_server_load_session_handler}
\title{Internal session loading handler for Proteomics DA module}
\usage{
da_server_load_session_handler(
  input,
  output,
  session,
  da_data,
  workflow_data,
  experiment_paths
)
}
\description{
Internal session loading handler for Proteomics DA module
}
\keyword{internal}
