% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_s4_objects.R
\docType{class}
\name{de_results_for_enrichment-class}
\alias{de_results_for_enrichment-class}
\title{de_results_for_enrichment S4 Class}
\description{
An S4 class to store differential expression results formatted for
enrichment analysis. Contains contrast definitions, DE data, and
experimental design information.
}
\section{Slots}{

\describe{
\item{\code{contrasts}}{A tibble containing contrast information}

\item{\code{de_data}}{A list of DE results data frames}

\item{\code{design_matrix}}{A data frame containing the design matrix}
}}

