% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_de_analysis_wrapper.R
\name{differentialExpressionAnalysisHelper}
\alias{differentialExpressionAnalysisHelper}
\title{Core Differential Expression Analysis Helper}
\usage{
differentialExpressionAnalysisHelper(
  theObject,
  contrasts_tbl = NULL,
  formula_string = NULL,
  group_id = NULL,
  de_q_val_thresh = NULL,
  treat_lfc_cutoff = NULL,
  eBayes_trend = NULL,
  eBayes_robust = NULL,
  args_group_pattern = NULL,
  args_row_id = NULL,
  qvalue_column = "fdr_qvalue",
  raw_pvalue_column = "raw_pvalue"
)
}
\description{
Core Differential Expression Analysis Helper
}
