% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_de.R
\name{differentialExpressionAppletModule}
\alias{differentialExpressionAppletModule}
\alias{mod_prot_de_ui}
\alias{mod_prot_de_server}
\title{differentialExpressionAppletModule}
\usage{
mod_prot_de_ui(id)

mod_prot_de_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label,
  selected_tab = NULL
)
}
\description{
A Shiny module for the Differential Expression step of the proteomics
workflow. Handles limma-based statistical analysis and visualization of results
across user-defined contrasts.
}
