% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_uniprot_id_helpers.R
\name{.cleanProteinIds}
\alias{.cleanProteinIds}
\title{Clean Protein IDs}
\usage{
.cleanProteinIds(string)
}
\arguments{
\item{string}{Character vector of protein IDs}
}
\value{
Cleaned protein IDs with isoform numbers removed
}
\description{
Remove isoform numbers from protein IDs (wrapper for cleanIsoformNumber)
}
