% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_multiomics_enrich_stringdb_api.R
\name{downloadStringDBResultsFile}
\alias{downloadStringDBResultsFile}
\title{Download STRING DB Results File}
\usage{
downloadStringDBResultsFile(download_url)
}
\arguments{
\item{download_url}{Character string: The direct URL to the results file.}
}
\value{
A data frame containing the parsed results from the URL.
Returns \code{NULL} if the download or parsing fails.
Messages about the download status are printed to the console.
}
\description{
Downloads a results file (typically TSV) from a given URL provided by the STRING API
and parses it into an R data frame.
}
\examples{
\dontrun{
# This function is typically called by retrieveStringDBEnrichmentResults
# but can be used standalone if you have a direct download URL.
# example_download_url <- "https://version-12-0.string-db.org/api/tsv/downloadenrichmentresults?networkId=bNEXfEymvDsZ"
# results_df <- downloadStringDBResultsFile(download_url = example_download_url)
# if (!is.null(results_df)) {
#   print(head(results_df))
# }
}
}
