% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_enrichment_plotting_helpers.R
\name{enrichedGoTermBarPlot}
\alias{enrichedGoTermBarPlot}
\title{Draw list of functional enrichment heatmaps}
\usage{
enrichedGoTermBarPlot(
  input_table,
  output_dir,
  analysis_type = "GO",
  file_suffix,
  width = 10,
  height = 7
)
}
\arguments{
\item{input_table, output_dir, analysis_type, file_suffix, width, height}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
given input table, draw a bar plot representing the GO enrichment results.
The height of each bar represents the negative log (base 10) q-values of the query proteins.
}
