% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_s4_objects.R
\docType{class}
\name{EnrichmentResults-class}
\alias{EnrichmentResults-class}
\title{EnrichmentResults S4 Class}
\description{
An S4 class to store enrichment analysis results including
enrichment data, plots, and summaries.
}
\section{Slots}{

\describe{
\item{\code{contrasts}}{A tibble containing contrast information}

\item{\code{enrichment_data}}{A list of enrichment results}

\item{\code{enrichment_plots}}{A list of gostplot objects}

\item{\code{enrichment_plotly}}{A list of interactive plotly objects}

\item{\code{enrichment_summaries}}{A list of summary data}
}}

