% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_import_organisms.R
\name{extractOrganismsFromFasta}
\alias{extractOrganismsFromFasta}
\title{Extract Organism Information from FASTA File}
\usage{
extractOrganismsFromFasta(fasta_file_path)
}
\arguments{
\item{fasta_file_path}{Path to the FASTA file}
}
\value{
A tibble with columns: uniprot_acc, organism_name, taxon_id
}
\description{
Parses FASTA file headers to extract organism name (OS) and
taxonomy ID (OX) for each protein entry. Used for analyzing mixed-species
FASTA databases.
}
