% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_s4_objects.R
\name{filtering_progress}
\alias{filtering_progress}
\title{Initialize a new FilteringProgress object}
\usage{
filtering_progress
}
\value{
A new FilteringProgress object
}
\description{
Creates a new FilteringProgress object with empty slots to track protein
filtering progress.
}
\examples{
filtering_progress <- new("FilteringProgress",
  steps = character(),
  proteins = numeric(),
  total_peptides = numeric(),
  peptides_per_protein = list(),
  proteins_per_run = list(),
  peptides_per_run = list()
)

}
