% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_progress_helpers.R
\docType{class}
\name{FilteringProgressLipidomics-class}
\alias{FilteringProgressLipidomics-class}
\title{FilteringProgressLipidomics S4 Class}
\description{
An S4 class to track lipidomics data filtering progress through QC steps.
}
