% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_qc_progress_helpers.R
\docType{class}
\name{FilteringProgressMetabolomics-class}
\alias{FilteringProgressMetabolomics-class}
\title{FilteringProgressMetabolomics S4 Class}
\description{
An S4 class to track metabolomics data filtering progress through QC steps.
}
