% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_peptide_methods.R
\name{filterMinNumPeptidesPerProtein,PeptideQuantitativeData-method}
\alias{filterMinNumPeptidesPerProtein,PeptideQuantitativeData-method}
\title{Filter the proteins based on the number of peptides and peptidoforms}
\usage{
\S4method{filterMinNumPeptidesPerProtein}{PeptideQuantitativeData}(theObject, ...)
}
\arguments{
\item{theObject}{Object of class PeptideQuantitativeData}

\item{...}{Filtering overrides, including \code{num_peptides_per_protein_thresh},
\code{num_peptidoforms_per_protein_thresh}, and \code{core_utilisation}.}
}
\description{
Keep protein groups with at least one distinct stripped peptide
and at least two distinct peptidoforms by default.
}
