% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_correlation_helpers.R
\name{filterSamplesByPeptideCorrelationThreshold}
\alias{filterSamplesByPeptideCorrelationThreshold}
\title{Filter Samples by Peptide Correlation Threshold}
\usage{
filterSamplesByPeptideCorrelationThreshold(
  pearson_correlation_per_pair,
  peptide_keep_samples_with_min_num_peptides,
  min_pearson_correlation_threshold = 0.75,
  filename_column_x = ms_filename.x,
  filename_column_y = ms_filename.y,
  correlation_column = pearson_correlation,
  filename_id_column = "Run"
)
}
\arguments{
\item{pearson_correlation_per_pair}{A data frame with the following columns: 1. ID of technical replicate group, 2. sample file name X, 3. sample file name Y, 4. Pearson correlation of the abundances of peptides between sample X and Y.}

\item{peptide_keep_samples_with_min_num_peptides}{A data frame with the following columns: 1. Sample file name or Run name, 2. Protein IDs, 3. Stripped peptide sequence, 4. Normalised peptide abundances}

\item{min_pearson_correlation_threshold}{Minimum pearson correlation for a pair of files to be considered to be consistent and kept for further analysis}

\item{filename_column_x}{Name of column containing the sample file name X (for a pair of sample in the same technical replicate group). Tidyverse column header format, not a string.}

\item{filename_column_y}{Name of column containing the sample file name Y (for a pair of sample in the same technical replicate group). Tidyverse column header format, not a string.}

\item{correlation_column}{Name of column containing the Pearson's correlation score between Sample X and Y. Tidyverse column header format, not a string.}

\item{filename_id_column}{A string indicating the name of column that contains the sample ID or Run ID in the data frame \code{peptide_keep_samples_with_min_num_peptides}.}
}
\value{
A table without samples that are poorly correlated with the rest of the samples in the technical replicate group. Contains the following columns: 1. Sample file name or Run name, 2. Protein IDs, 3. Stripped peptide sequence, 4. Normalised peptide abundances
}
\description{
Remove samples which is correlated with any technical replicate samples
}
