% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoproteomics_helper.R
\name{formatPhosphositePosition}
\alias{formatPhosphositePosition}
\title{formatPhosphositePosition}
\usage{
formatPhosphositePosition(gene_symbol, position, residue, delim = ";")
}
\arguments{
\item{gene_symbol}{The gene symbol as a string}

\item{position}{The positions of the phosphorylation sites as a string, multiple sites are separated by the delimiter (e.g. delim parameter). Multiple positions are in the same corresponding order as the residue parameter.}

\item{residue}{The amino acid residue where the phosphorylation is at, multiple residues are separate by the delimiter (e.g. delim paramter). Multiple residues are in the same corresponding order as the position parameter.}
}
\value{
A string with the gene symbol, followed by a semi-colon, and the amino acid residue and the position, multiple residue and positions separated by commas (e.g. YFG1;S199,T203 )
}
\description{
formatPhosphositePosition
}
