% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_objects.R, R/func_prot_de.R
\name{generateVolcanoPlotGlimma}
\alias{generateVolcanoPlotGlimma}
\title{Generate Interactive Volcano Plot using Glimma}
\usage{
generateVolcanoPlotGlimma(
  de_results_list,
  selected_contrast = NULL,
  uniprot_tbl = NULL,
  args_row_id = "uniprot_acc",
  fdr_column = "fdr_qvalue",
  raw_p_value_column = "raw_pvalue",
  log2fc_column = "log2FC",
  de_q_val_thresh = 0.05,
  uniprot_id_column = "Entry",
  gene_names_column = "gene_names",
  display_columns = c("best_uniprot_acc")
)

generateVolcanoPlotGlimma(
  de_results_list,
  selected_contrast = NULL,
  uniprot_tbl = NULL,
  args_row_id = "uniprot_acc",
  fdr_column = "fdr_qvalue",
  raw_p_value_column = "raw_pvalue",
  log2fc_column = "log2FC",
  de_q_val_thresh = 0.05,
  uniprot_id_column = "Entry",
  gene_names_column = "gene_names",
  display_columns = c("best_uniprot_acc")
)
}
\description{
Generate Interactive Volcano Plot using Glimma

Generate Interactive Volcano Plot using Glimma
}
