% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_palette_theme_helpers.R
\name{getCmriMachineColour}
\alias{getCmriMachineColour}
\title{Standard colours for CMRI mass spectrometers}
\usage{
getCmriMachineColour()
}
\value{
A named character vector of colours for machines M01 through M08.
}
\description{
Standard colours for CMRI mass spectrometers
}
