% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_palette_theme_helpers.R
\name{getContinousColourRules}
\alias{getContinousColourRules}
\title{getContinousColourRules}
\usage{
getContinousColourRules(
  metadata_tbl,
  metadata_column_labels,
  metadata_column_selected,
  continous_scale_columns,
  na_colour = "white"
)
}
\arguments{
\item{metadata_tbl, metadata_column_labels, metadata_column_selected, continous_scale_columns, na_colour}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
getContinousColourRules
}
