% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{getGlimmaVolcanoPhosphoproteomics}
\alias{getGlimmaVolcanoPhosphoproteomics}
\title{getGlimmaVolcanoPhosphoproteomics}
\usage{
getGlimmaVolcanoPhosphoproteomics(
  r_obj,
  coef,
  volcano_plot_tab,
  sites_id_column = sites_id,
  sites_id_display_column = sites_id_short,
  display_columns = c("sequence", "PROTEIN_NAMES"),
  additional_annotations = NULL,
  additional_annotations_join_column = NULL,
  counts_tbl = NULL,
  output_dir
)
}
\arguments{
\item{r_obj}{Output from ebFit object of limma package}

\item{coef}{An integer specifying the position in the list of coefficients (e.g. name of contrast) for which to print the volcano plot for}

\item{volcano_plot_tab}{A table containing the list of uniprot_acc and the matching gene_name.}

\item{sites_id_column}{The name of the column in the 'volcano_plot_tab' table that contains the list of uniprot accessions (in tidyverse format).}

\item{sites_id_display_column}{The name of the column in the 'volcano_plot_tab' table that contains the list of gene names (in tidyverse format).}

\item{display_columns}{The name of the columns from input \code{volcano_plot_tab} that will be included in the mouseover tooltips and table}

\item{output_dir}{The output directory in which the HTML files containing the interactive plotly volcano plot will be saved.}
}
\description{
Create an interactive plotly volcano plot for Phosphoproteomics data
}
