% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{getGlimmaVolcanoProteomics}
\alias{getGlimmaVolcanoProteomics}
\title{getGlimmaVolcanoProteomics}
\usage{
getGlimmaVolcanoProteomics(
  r_obj,
  coef,
  volcano_plot_tab,
  uniprot_column = best_uniprot_acc,
  gene_name_column = gene_name,
  display_columns = c("PROTEIN_NAMES"),
  additional_annotations = NULL,
  additional_annotations_join_column = NULL,
  counts_tbl = NULL,
  groups = NULL,
  output_dir
)
}
\arguments{
\item{r_obj}{Output from ebFit object of limma package}

\item{coef}{An integer specifying the position in the list of coefficients (e.g. name of contrast) for which to print the volcano plot for}

\item{volcano_plot_tab}{A table containing the list of uniprot_acc and the matching gene_name.}

\item{uniprot_column}{The name of the column in the 'volcano_plot_tab' table that contains the list of uniprot accessions (in tidyverse format).}

\item{gene_name_column}{The name of the column in the 'volcano_plot_tab' table that contains the list of gene names (in tidyverse format).}

\item{counts_tbl}{A table containing the intensity data for the proteins.}

\item{output_dir}{The output directory in which the HTML files containing the interactive plotly volcano plot will be saved.}
}
\description{
Create an interactive plotly volcano plot for Proteomics data
}
