% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinVsSamplesS4Objects.R
\name{getLowCoefficientOfVariationProteins}
\alias{getLowCoefficientOfVariationProteins}
\title{Get Low Coefficient of Variation Proteins}
\usage{
getLowCoefficientOfVariationProteins(
  theObject,
  percentage_as_neg_ctrl = NULL,
  num_neg_ctrl = NULL
)
}
\description{
Sort proteins by their coefficient of variation and take the top N with lowest coefficient of variation
}
