% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{getOneContinousPalette}
\alias{getOneContinousPalette}
\title{getOneContinousPalette}
\usage{
getOneContinousPalette(
  metadata_tbl,
  column_name,
  palette_name,
  na_colour = "white"
)
}
\description{
getOneContinousPalette
}
