% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/func_general_plotting_heatmap_support_helpers.R
\name{getProteinsHeatMap}
\alias{getProteinsHeatMap}
\title{get protein intensity heatmap}
\usage{
getProteinsHeatMap(
  protein_matrix,
  metadata_tbl,
  is_HEK_column = is_HEK,
  metadata_column_selected,
  metadata_column_labels,
  colour_rules,
  columns_to_exclude,
  core_utilisation_samples = TRUE,
  sort_by_sample_id = TRUE,
  sample_id_column = Run,
  use_raster = TRUE,
  raster_device = "CairoTIFF",
  heatmap_legend_param = list(title = "Intensity")
)
}
\arguments{
\item{protein_matrix, metadata_tbl, is_HEK_column, metadata_column_selected, metadata_column_labels, colour_rules, columns_to_exclude, core_utilisation_samples, sort_by_sample_id, sample_id_column, use_raster, raster_device, heatmap_legend_param}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Generates a heatmap of protein intensities
}
