% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_diagnostic_helpers.R
\name{getSamplesCorrelationMatrix}
\alias{getSamplesCorrelationMatrix}
\title{getSamplesCorrelationMatrix}
\usage{
getSamplesCorrelationMatrix(
  input_table,
  metadata_tbl,
  is_HEK_column = is_HEK,
  use = "pairwise.complete.obs",
  method = "pearson"
)
}
\arguments{
\item{input_table}{Table with samples as columns and peptides as rows. Contains the log peptide intensity values.}
}
\description{
Calculate the Pearson's correlation score between sample
}
