% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_phospho_annotation.R,
%   R/phosphoproteomics_helper.R
\name{getUniprotAccRankFromSitesId}
\alias{getUniprotAccRankFromSitesId}
\title{Get the ranking of the uniprot_acc within the sites_id}
\usage{
getUniprotAccRankFromSitesId(input_table, uniprot_acc, sites_id)

getUniprotAccRankFromSitesId(input_table, uniprot_acc, sites_id)
}
\arguments{
\item{input_table, }{an input table with sites_id and uniprot_acc columns}

\item{uniprot_acc, }{name of Uniprot accession column tidyverse style column name input}

\item{sites_id, }{name of sites_id column, in tidyverse style column name}
}
\description{
Given an input table with sites_id and uniprot_acc columns, work out the ranking of the uniprot_acc within the sites_id

Given an input table with sites_id and uniprot_acc columns, work out the ranking of the uniprot_acc within the sites_id
}
