% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_uniprot_id_helpers.R
\name{getUniprotRegexPatterns}
\alias{getUniprotRegexPatterns}
\title{Get UniProt Regex Patterns}
\usage{
getUniprotRegexPatterns()
}
\value{
A list with patterns: isoform, decoy, contaminant, delimiter
}
\description{
Returns a named list of regex patterns used for UniProt ID management.
}
