% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_go_helpers.R
\name{goIdToTerm}
\alias{goIdToTerm}
\title{Convert a list of Gene Ontology IDs to their respective human readable name (e.g. GO term).}
\usage{
goIdToTerm(go_string, sep = "; ", goterms, gotypes)
}
\arguments{
\item{go_string}{A string consisting of a list of Gene Ontology ID, separated by a delimiter}

\item{sep}{Delimiter separating GO IDs in \code{go_string}.}

\item{goterms}{Output from running \code{goterms <- Term(GOTERM)} from the GO.db library.}

\item{gotypes}{Output from running \code{gotypes <- Ontology(GOTERM)} from the GO.db library.}
}
\value{
A table with three columns. go_biological_process, go_celluar_compartment, and go_molecular_function. Each column is a list of gene ontology terms, separated by '; '.
}
\description{
Convert a list of Gene Ontology IDs to their respective human readable name (e.g. GO term).
}
\examples{
go_string <- "GO:0016021; GO:0045087"
goterms <- c(
  "GO:0016021" = "integral component of membrane",
  "GO:0045087" = "innate immune response"
)
gotypes <- c("GO:0016021" = "CC", "GO:0045087" = "BP")
goIdToTerm(go_string, goterms = goterms, gotypes = gotypes)
}
