% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_import_readers.R
\name{importFragPipeData}
\alias{importFragPipeData}
\title{Import FragPipe Data}
\usage{
importFragPipeData(filepath, use_maxlfq = TRUE)
}
\arguments{
\item{filepath}{Path to FragPipe output file}

\item{use_maxlfq}{Whether to use MaxLFQ intensities}
}
\value{
List with data, data_type, and column_mapping
}
\description{
Import FragPipe Data
}
