% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_import_readers.R
\name{importMaxQuantData}
\alias{importMaxQuantData}
\title{Import MaxQuant Data}
\usage{
importMaxQuantData(filepath, use_lfq = TRUE, filter_contaminants = TRUE)
}
\arguments{
\item{filepath}{Path to MaxQuant proteinGroups.txt file}

\item{use_lfq}{Whether to use LFQ intensities}

\item{filter_contaminants}{Whether to filter contaminants}
}
\value{
List with data, data_type, and column_mapping
}
\description{
Import MaxQuant Data
}
