% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_import_readers.R
\name{importSpectronautData}
\alias{importSpectronautData}
\title{Import Spectronaut Data}
\usage{
importSpectronautData(filepath, quantity_type = "pg")
}
\arguments{
\item{filepath}{Path to Spectronaut report file}

\item{quantity_type}{Either "pg" for protein or "pep" for peptide quantities}
}
\value{
List with data, data_type, and column_mapping
}
\description{
Import Spectronaut Data
}
