% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_normalization_methods.R
\name{lipidIntensityFiltering,LipidomicsAssayData-method}
\alias{lipidIntensityFiltering,LipidomicsAssayData-method}
\title{Lipid Intensity Filtering Method for LipidomicsAssayData}
\usage{
\S4method{lipidIntensityFiltering}{LipidomicsAssayData}(
  theObject,
  lipids_intensity_cutoff_percentile = NULL,
  lipids_proportion_of_samples_below_cutoff = NULL
)
}
\arguments{
\item{theObject}{A LipidomicsAssayData object.}

\item{lipids_intensity_cutoff_percentile}{See generic definition.}

\item{lipids_proportion_of_samples_below_cutoff}{See generic definition.}
}
\value{
An updated LipidomicsAssayData object.
}
\description{
Filters lipids in \emph{all} assays of a LipidomicsAssayData object.
It removes lipids that have intensities below a certain percentile threshold
in a proportion of samples exceeding a defined cutoff. The threshold is calculated
independently for each assay.
}
\section{Functions}{
\itemize{
\item \code{lipidIntensityFiltering(LipidomicsAssayData)}: Method for LipidomicsAssayData

}}
