% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_pept_norm.R
\name{log2Transformation}
\alias{log2Transformation}
\title{Log2 Transformation with Pseudo-count}
\usage{
log2Transformation(input_matrix)
}
\arguments{
\item{input_matrix}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Log 2 transformation with pseudo count
}
