% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_pept_s4_norm_methods.R
\name{log2TransformPeptideMatrix,PeptideQuantitativeData-method}
\alias{log2TransformPeptideMatrix,PeptideQuantitativeData-method}
\title{Log2 Transform Peptide Matrix}
\usage{
\S4method{log2TransformPeptideMatrix}{PeptideQuantitativeData}(theObject)
}
\arguments{
\item{theObject}{A PeptideQuantitativeData object}
}
\value{
PeptideQuantitativeData object with log2 transformed data
}
\description{
Transforms raw peptide intensity values to log2 scale for downstream normalization and RUV analysis.
This should be called after calcPeptideMatrix() and before normaliseBetweenSamples().
}
