% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_qc_methods.R
\name{metaboliteIntensityFiltering,MetaboliteAssayData-method}
\alias{metaboliteIntensityFiltering,MetaboliteAssayData-method}
\title{Metabolite Intensity Filtering Method for MetaboliteAssayData}
\usage{
\S4method{metaboliteIntensityFiltering}{MetaboliteAssayData}(
  theObject,
  metabolites_intensity_cutoff_percentile = NULL,
  metabolites_proportion_of_samples_below_cutoff = NULL
)

\S4method{metaboliteIntensityFiltering}{MetaboliteAssayData}(
  theObject,
  metabolites_intensity_cutoff_percentile = NULL,
  metabolites_proportion_of_samples_below_cutoff = NULL
)
}
\arguments{
\item{theObject}{A MetaboliteAssayData object.}

\item{metabolites_intensity_cutoff_percentile}{See generic definition.}

\item{metabolites_proportion_of_samples_below_cutoff}{See generic definition.}
}
\value{
An updated MetaboliteAssayData object.

An updated MetaboliteAssayData object.
}
\description{
Filters metabolites in \emph{all} assays of a MetaboliteAssayData object.
It removes metabolites that have intensities below a certain percentile threshold
in a proportion of samples exceeding a defined cutoff. The threshold is calculated
independently for each assay.

Filters metabolites in \emph{all} assays of a MetaboliteAssayData object.
It removes metabolites that have intensities below a certain percentile threshold
in a proportion of samples exceeding a defined cutoff. The threshold is calculated
independently for each assay.
}
\section{Functions}{
\itemize{
\item \code{metaboliteIntensityFiltering(MetaboliteAssayData)}: Method for MetaboliteAssayData

\item \code{metaboliteIntensityFiltering(MetaboliteAssayData)}: Method for MetaboliteAssayData

}}
