% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lipid_da_ui.R, R/mod_lipid_da_server.R,
%   R/mod_lipid_da.R
\name{mod_lipid_da_ui}
\alias{mod_lipid_da_ui}
\alias{mod_lipid_da_server}
\alias{mod_lipid_da}
\title{Lipidomics Differential Analysis Module}
\usage{
mod_lipid_da_ui(id)

mod_lipid_da_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label
)
}
\arguments{
\item{id, workflow_data, experiment_paths, omic_type, experiment_label}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for performing per-assay differential abundance
analysis on lipidomics data using limma.
}
