% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lipid_de.R
\name{mod_lipid_de}
\alias{mod_lipid_de}
\alias{mod_lipid_de_ui}
\alias{mod_lipid_de_server}
\title{Lipidomics Differential Analysis Module}
\usage{
mod_lipid_de_ui(id)

mod_lipid_de_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for performing per-assay differential abundance
analysis on lipidomics data using limma.
}
