% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metab_de.R
\name{mod_metab_de}
\alias{mod_metab_de}
\alias{mod_metab_de_ui}
\alias{mod_metab_de_server}
\title{Metabolomics Differential Analysis Module}
\usage{
mod_metab_de_ui(id)

mod_metab_de_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for performing per-assay differential abundance
analysis on metabolomics data using limma.
}
