% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metab_qc_duplicates_ui.R,
%   R/mod_metab_qc_duplicates_server.R
\name{mod_metab_qc_duplicates}
\alias{mod_metab_qc_duplicates}
\alias{mod_metab_qc_duplicates_ui}
\alias{mod_metab_qc_duplicates_server}
\title{Metabolite Duplicate Resolution Module}
\usage{
mod_metab_qc_duplicates_ui(id)

mod_metab_qc_duplicates_server(id, workflow_data, omic_type, experiment_label)
}
\arguments{
\item{id, workflow_data, omic_type, experiment_label}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for identifying and resolving duplicate metabolite features.
}
