% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_impute.R
\name{mod_prot_qc_peptide_impute}
\alias{mod_prot_qc_peptide_impute}
\alias{mod_prot_qc_peptide_impute_ui}
\alias{.peptideImputationAuditSummary}
\title{Peptide Imputation Module}
\usage{
mod_prot_qc_peptide_impute_ui(id)

.peptideImputationAuditSummary(
  imputed_s4,
  proportion_missing_values,
  exclusion_column = NULL
)
}
\arguments{
\item{id, imputed_s4, proportion_missing_values, exclusion_column}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying missing value imputation.
}
