% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_intensity.R
\name{mod_prot_qc_peptide_intensity}
\alias{mod_prot_qc_peptide_intensity}
\alias{mod_prot_qc_peptide_intensity_ui}
\alias{buildPeptideIntensityThresholdPreview}
\title{Peptide Intensity Filter Module}
\usage{
mod_prot_qc_peptide_intensity_ui(id)

buildPeptideIntensityThresholdPreview(
  workflowData,
  minRepsPerGroup,
  minGroups,
  updateMissingValueParametersFn = updateMissingValueParameters
)
}
\arguments{
\item{id, workflowData, minRepsPerGroup, minGroups, updateMissingValueParametersFn}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying peptide intensity filtering.
}
