% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_protein.R
\name{mod_prot_qc_peptide_protein}
\alias{mod_prot_qc_peptide_protein}
\alias{mod_prot_qc_peptide_protein_ui}
\alias{runProteinPeptideApplyStep}
\title{Protein Peptide Count Filter Module}
\usage{
mod_prot_qc_peptide_protein_ui(id)

runProteinPeptideApplyStep(
  workflowData,
  minPeptidesPerProtein,
  minPeptidoformsPerProtein,
  updateConfigParameterFn = updateConfigParameter,
  filterMinNumPeptidesPerProteinFn = filterMinNumPeptidesPerProtein,
  logInfoFn = logger::log_info,
  nowFn = Sys.time
)
}
\arguments{
\item{id, workflowData, minPeptidesPerProtein, minPeptidoformsPerProtein, updateConfigParameterFn, filterMinNumPeptidesPerProteinFn, logInfoFn, nowFn}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying the protein peptide count filter.
}
