% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_qvalue.R
\name{mod_prot_qc_peptide_qvalue}
\alias{mod_prot_qc_peptide_qvalue}
\alias{mod_prot_qc_peptide_qvalue_ui}
\alias{runPeptideQvalueApplyStep}
\title{Peptide Q-Value Filter Module}
\usage{
mod_prot_qc_peptide_qvalue_ui(id)

runPeptideQvalueApplyStep(
  workflowData,
  qvalueThreshold,
  globalQvalueThreshold,
  proteotypicOnly,
  updateConfigParameterFn = updateConfigParameter,
  qvalueFilterFn = srlQvalueProteotypicPeptideClean,
  logInfoFn = logger::log_info,
  logWarnFn = logger::log_warn,
  nowFn = Sys.time,
  globalPGQvalueThreshold = .diaNnIdentificationQvalueCutoff,
  confidenceProvenanceMode = "diann_main_report"
)
}
\arguments{
\item{id, workflowData, qvalueThreshold, globalQvalueThreshold, proteotypicOnly, updateConfigParameterFn, qvalueFilterFn, logInfoFn, logWarnFn, nowFn, globalPGQvalueThreshold, confidenceProvenanceMode}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying Q-value and proteotypic peptide filters.
}
