% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_replicate.R
\name{mod_prot_qc_peptide_replicate}
\alias{mod_prot_qc_peptide_replicate}
\alias{mod_prot_qc_peptide_replicate_ui}
\alias{runPeptideReplicateRevertStep}
\title{Peptide Replicate Filter Module}
\usage{
mod_prot_qc_peptide_replicate_ui(id)

runPeptideReplicateRevertStep(workflowData, logInfoFn = logger::log_info)
}
\arguments{
\item{id, workflowData, logInfoFn}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying the replicate filter.
}
