% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_rollup.R
\name{mod_prot_qc_peptide_rollup}
\alias{mod_prot_qc_peptide_rollup}
\alias{mod_prot_qc_peptide_rollup_ui}
\alias{.peptideRollupSummaryFromObject}
\title{Peptide Precursor Rollup Module}
\usage{
mod_prot_qc_peptide_rollup_ui(id)

.peptideRollupSummaryFromObject(object)
}
\arguments{
\item{id, object}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying precursor to peptide rollup.
}
