% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_sample.R
\name{mod_prot_qc_peptide_sample}
\alias{mod_prot_qc_peptide_sample}
\alias{mod_prot_qc_peptide_sample_ui}
\alias{runPeptideSampleApplyStep}
\title{Peptide Sample Filter Module}
\usage{
mod_prot_qc_peptide_sample_ui(id)

runPeptideSampleApplyStep(
  workflowData,
  minPeptidesPerSample,
  updateConfigParameterFn = updateConfigParameter,
  filterMinNumPeptidesPerSampleFn = filterMinNumPeptidesPerSample,
  logInfoFn = logger::log_info,
  nowFn = Sys.time
)
}
\arguments{
\item{id, workflowData, minPeptidesPerSample, updateConfigParameterFn, filterMinNumPeptidesPerSampleFn, logInfoFn, nowFn}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying the sample quality filter.
}
