% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_cleanup.R
\name{mod_prot_qc_protein_cleanup}
\alias{mod_prot_qc_protein_cleanup}
\alias{mod_prot_qc_protein_cleanup_ui}
\alias{.proteinCleanupActiveKey}
\title{Protein Accession Cleanup Module}
\usage{
mod_prot_qc_protein_cleanup_ui(id)

.proteinCleanupActiveKey(proteinObject)
}
\arguments{
\item{id, proteinObject}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for performing protein accession cleanup.
}
