% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_dedup.R
\name{mod_prot_qc_protein_dedup}
\alias{mod_prot_qc_protein_dedup}
\alias{mod_prot_qc_protein_dedup_ui}
\alias{runProteinDuplicateRemovalStep}
\title{Protein Duplicate Removal Module}
\usage{
mod_prot_qc_protein_dedup_ui(id)

runProteinDuplicateRemovalStep(
  workflowData,
  aggregationMethod,
  aggregationResolverFn = base::get,
  logInfoFn = logger::log_info
)
}
\arguments{
\item{id, workflowData, aggregationMethod, aggregationResolverFn, logInfoFn}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for removing duplicate proteins.
}
