% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_intensity.R
\name{mod_prot_qc_protein_intensity}
\alias{mod_prot_qc_protein_intensity}
\alias{mod_prot_qc_protein_intensity_ui}
\alias{runProteinIntensityFilterApplyStep}
\title{Protein Intensity Filter Module}
\usage{
mod_prot_qc_protein_intensity_ui(id)

runProteinIntensityFilterApplyStep(
  workflowData,
  useStrictMode,
  minRepsPerGroup,
  minGroups,
  intensityCutoffPercentile,
  updateConfigParameterFn = updateConfigParameter,
  updateMissingValueParametersFn = updateMissingValueParameters,
  removeRowsWithMissingValuesPercentFn = removeRowsWithMissingValuesPercent,
  logInfoFn = logger::log_info,
  nowFn = Sys.time
)
}
\arguments{
\item{id, workflowData, useStrictMode, minRepsPerGroup, minGroups, intensityCutoffPercentile, updateConfigParameterFn, updateMissingValueParametersFn, removeRowsWithMissingValuesPercentFn, logInfoFn, nowFn}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
A Shiny module for applying protein intensity and missing value filters.
}
