% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_proteomics.R
\name{mod_proteomics_ui}
\alias{mod_proteomics_ui}
\title{Proteomics Workflow UI Module}
\usage{
mod_proteomics_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Main UI for the proteomics workflow interface
}
