% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_legacy_api_docs.R
\name{MultiScholaR-legacy-api}
\alias{MultiScholaR-legacy-api}
\alias{\%||\%}
\alias{add_node_label}
\alias{addColumnsToEvidenceTbl}
\alias{addPeptideStartAndEnd}
\alias{addPhosphositesPositionsString}
\alias{addXMerStrings}
\alias{allPhosphositesPivotLonger}
\alias{allPhosphositesPivotWider}
\alias{averageTechReps}
\alias{batchQueryEvidence}
\alias{batchQueryEvidenceGeneId}
\alias{buildAnnotationIdToAnnotationNameDictionary}
\alias{buildOneProteinToAnnotationList}
\alias{calcHtSize}
\alias{calcPeptideMatrix}
\alias{calcPeptideMatrix,PeptideQuantitativeData-method}
\alias{calculateLipidCVs}
\alias{calculateLipidMissingness}
\alias{calculateLipidSumIntensityPerSample}
\alias{calculateMetaboliteCVs}
\alias{calculateMissingness}
\alias{calculateMissingValuesPerProteinFishersTest}
\alias{calculatePercentMissingPerProtein}
\alias{calculateSumIntensityPerSample}
\alias{calculateTotalUniqueLipidsAcrossAssays}
\alias{calculateTotalUniqueMetabolitesAcrossAssays}
\alias{chooseBestPhosphositeAccession}
\alias{chooseBestProteinAccession}
\alias{chooseBestProteinAccession,PeptideQuantitativeData-method}
\alias{chooseBestProteinAccessionSumDuplicates}
\alias{cleanDesignMatrix}
\alias{cleanDesignMatrix,ProteinQuantitativeData-method}
\alias{cleanDesignMatrixCleanCategories}
\alias{cleanDesignMatrixCleanCategoriesMap}
\alias{cleanDesignMatrixCreateEachVersusAllColumns}
\alias{cleanDesignMatrixPeptide}
\alias{cleanDesignMatrixPeptide,PeptideQuantitativeData-method}
\alias{cleanDuplicatesEnrichment}
\alias{clusterPathways}
\alias{cnetplotEdited}
\alias{compareTwoPeptideDataObjects}
\alias{compareUmapComponentsPairs}
\alias{convertIdToAnnotation}
\alias{convertKeyToAttribute}
\alias{convertProteinAccToGeneSymbol}
\alias{count_num_peptides}
\alias{count_num_proteins}
\alias{count_num_samples}
\alias{countLipidsPerSample}
\alias{countMetabolitesPerSample}
\alias{countStatDaGenesHelper}
\alias{countUniqueLipids}
\alias{countUniqueMetabolites}
\alias{createDaResultsLongFormat}
\alias{createDirectoryIfNotExists}
\alias{createDirIfNotExists}
\alias{createGridQC}
\alias{createGridQC,GridPlotData-method}
\alias{createIdToAttributeHash}
\alias{createOutputDir}
\alias{daAnalysisWrapperFunction}
\alias{differentialAbundanceAnalysis}
\alias{differentialAbundanceAnalysis,list-method}
\alias{differentialAbundanceAnalysis,ProteinQuantitativeData-method}
\alias{differentialAbundanceAnalysisHelper}
\alias{differentialAbundanceAnalysisHelper,LipidomicsAssayData-method}
\alias{differentialAbundanceAnalysisHelper,MetaboliteAssayData-method}
\alias{differentialAbundanceAnalysisHelper,ProteinQuantitativeData-method}
\alias{download_uniprot_data}
\alias{drawListOfFunctionalEnrichmentHeatmaps}
\alias{drawListOfFunctionalEnrichmentHeatmapsScholar}
\alias{enrichedPathwayBarPlot}
\alias{evaluateBestMinMaxGeneSetSize}
\alias{extract_geneSets}
\alias{extractResults}
\alias{extractRuvResults}
\alias{filterByScoreAndGetSimilarPeptides}
\alias{filterMinNumPeptidesPerProtein}
\alias{filterMinNumPeptidesPerProtein,ProteinQuantitativeData-method}
\alias{filterMinNumPeptidesPerSample}
\alias{filterMinNumPeptidesPerSample,PeptideQuantitativeData-method}
\alias{filterPeptideAndExtractProbabilities}
\alias{filterResultsWithRevigo}
\alias{filterResultsWithRevigoScholar}
\alias{filterSamplesByLipidCorrelationThreshold}
\alias{filterSamplesByProteinCorrelationThreshold}
\alias{filterSamplesByProteinCorrelationThreshold,ProteinQuantitativeData-method}
\alias{findBestK}
\alias{findBestNegCtrlPercentagePeptides}
\alias{findSamplesPairBelowPeptideCorrelationThreshold}
\alias{generate_enrichment_plots}
\alias{generateLipidFilteringPlots}
\alias{generateMetaboliteFilteringPlots}
\alias{get_enrichplot_color}
\alias{get_ggrepel_segsize}
\alias{get_param_change_message}
\alias{getBestPosition}
\alias{getBestPositionFutureMap}
\alias{getCategoricalColourPalette}
\alias{getDaResultsLongFormat}
\alias{getDaResultsLongFormat,list-method}
\alias{getDaResultsWideFormat}
\alias{getDaResultsWideFormat,list-method}
\alias{getEnrichmentHeatmap}
\alias{getEnrichmentPlotly}
\alias{getEnrichmentResult}
\alias{getEnrichmentSummary}
\alias{getFastaFields}
\alias{getFilteringProgressLipidomics}
\alias{getFilteringProgressMetabolomics}
\alias{getFunctionName}
\alias{getFunctionNameSecondLevel}
\alias{getGlimmaVolcanoProteomicsWidget}
\alias{getInternalStandardMetrics}
\alias{getLipidInternalStandardMetrics}
\alias{getLipidQuantData}
\alias{getLowCoefficientOfVariationProteins}
\alias{getMaxProb}
\alias{getMaxProbFutureMap}
\alias{getMetaboliteQuantData}
\alias{getNegCtrlMetabAnova}
\alias{getNegCtrlProtAnova}
\alias{getNegCtrlProtAnova,ProteinQuantitativeData-method}
\alias{getNegCtrlProtAnovaPeptides}
\alias{getNegCtrlProtAnovaPeptides,PeptideQuantitativeData-method}
\alias{getOneContinousPalette}
\alias{getOneRlePlotData}
\alias{getPcaMatrix}
\alias{getPcaMatrix,ProteinQuantitativeData-method}
\alias{getRuvIIIReplicateMatrix}
\alias{getRuvIIIReplicateMatrix,ProteinQuantitativeData-method}
\alias{getUniprotAccToGeneSymbolDictionary}
\alias{getUniProtAnnotation}
\alias{getXMersList}
\alias{gg_save_logging}
\alias{GridPlotData-class}
\alias{groupParalogPeptides}
\alias{InitialiseGrid}
\alias{InitialiseGrid,ANY-method}
\alias{isArgumentDefined}
\alias{lipidIntensityFiltering}
\alias{list2df}
\alias{list2graph}
\alias{listifyTableByColumn}
\alias{log2TransformPeptideMatrix}
\alias{logTransformAssays}
\alias{matchAnnotations}
\alias{metaboliteIntensityFiltering}
\alias{node_add_alpha}
\alias{normaliseBetweenSamples}
\alias{normaliseUntransformedData}
\alias{outputDaAnalysisResults}
\alias{outputDaResultsAllContrasts}
\alias{outputDaResultsAllContrasts,ProteinQuantitativeData-method}
\alias{outputDeAnalysisResults}
\alias{parseList}
\alias{parseNumList}
\alias{parseString}
\alias{parseType}
\alias{pearsonCorForSamplePairs}
\alias{pearsonCorForSamplePairs,PeptideQuantitativeData-method}
\alias{peptideIntensityFiltering}
\alias{peptideIntensityFiltering,PeptideQuantitativeData-method}
\alias{peptideMissingValueImputation}
\alias{peptideMissingValueImputation,PeptideQuantitativeData-method}
\alias{peptideMissingValueImputationLimpa}
\alias{PeptideQuantitativeData-class}
\alias{PeptideQuantitativeDataDiann}
\alias{perform_enrichment}
\alias{plotDensity}
\alias{plotDensity,ggplot-method}
\alias{plotDensity,PeptideQuantitativeData-method}
\alias{plotDensity,ProteinQuantitativeData-method}
\alias{plotDensityList}
\alias{plotDensityList,ProteinQuantitativeData-method}
\alias{plotDensityOfProteinIntensityPerSample}
\alias{plotEnrichmentBarplot}
\alias{plotHistogramOfPercentMissingPerIndvidual}
\alias{plotInteractiveVolcano}
\alias{plotInteractiveVolcano,list-method}
\alias{plotNumSigDiffExpBarPlot}
\alias{plotNumSigDiffExpBarPlot,list-method}
\alias{plotPca}
\alias{plotPca,PeptideQuantitativeData-method}
\alias{plotPca,ProteinQuantitativeData-method}
\alias{plotPcaBox}
\alias{plotPcaGgpairs}
\alias{plotPcaList}
\alias{plotPcaList,ProteinQuantitativeData-method}
\alias{plotPcaListHelper}
\alias{plotPearson}
\alias{plotPearson,PeptideQuantitativeData-method}
\alias{plotPercentSamplesVsProteinQuantified}
\alias{plotRle}
\alias{plotRle,PeptideQuantitativeData-method}
\alias{plotRle,ProteinQuantitativeData-method}
\alias{plotRleList}
\alias{plotRleList,ProteinQuantitativeData-method}
\alias{plotRleQc}
\alias{plotVolcanoS4}
\alias{plotVolcanoS4,list-method}
\alias{preservePeptideNaValues}
\alias{preservePeptideNaValues,PeptideQuantitativeData,ProteinQuantitativeData-method}
\alias{preservePeptideNaValuesHelper}
\alias{processAndFilterData}
\alias{processMultisiteEvidence}
\alias{proteinIntensityFiltering}
\alias{proteinIntensityFiltering,ProteinQuantitativeData-method}
\alias{proteinMissingValueImputationLimpa}
\alias{proteinTechRepCorrelation}
\alias{proteinTechRepCorrelation,ProteinQuantitativeData-method}
\alias{queryRevigo}
\alias{readEnrichmentResultFiles}
\alias{removePeptidesOnlyInHek293}
\alias{removePeptidesWithMissingValuesPercent}
\alias{removePeptidesWithMissingValuesPercent,PeptideQuantitativeData-method}
\alias{removePeptidesWithOnlyOneReplicate}
\alias{removePeptidesWithOnlyOneReplicate,PeptideQuantitativeData-method}
\alias{removePeptidesWithoutAbundances}
\alias{removeProteinsWithOnlyOneReplicate}
\alias{removeProteinsWithOnlyOneReplicate,ProteinQuantitativeData-method}
\alias{removeRowsWithMissingValuesPercent}
\alias{removeRowsWithMissingValuesPercent,ProteinQuantitativeData-method}
\alias{RenderReport}
\alias{resolveDuplicateFeatures}
\alias{rlePcaPlotList}
\alias{rollUpPrecursorToPeptide}
\alias{rollUpPrecursorToPeptide,PeptideQuantitativeData-method}
\alias{runEnricher}
\alias{runGsea}
\alias{runOneGoEnrichmentInOutFunction}
\alias{ruvCancor}
\alias{ruvCancor,ProteinQuantitativeData-method}
\alias{ruvCancorFast}
\alias{ruvIII_C_Varying}
\alias{ruvIII_C_Varying,PeptideQuantitativeData-method}
\alias{ruvIII_C_Varying,ProteinQuantitativeData-method}
\alias{saveFilteredFunctionalEnrichmentTable}
\alias{saveListOfFunctionalEnrichmentHeatmaps}
\alias{saveListOfPdfs}
\alias{savePlotDensityList}
\alias{savePlotPcaList}
\alias{savePlotRleList}
\alias{set_enrichplot_color}
\alias{setArgsDefault}
\alias{setProteinData}
\alias{setProteinData,ProteinQuantitativeData-method}
\alias{sourceRmdFileSimple}
\alias{srlQvalueProteotypicPeptideClean}
\alias{srlQvalueProteotypicPeptideClean,PeptideQuantitativeData-method}
\alias{summarisePeptideObject}
\alias{summarize_enrichment}
\alias{summarizeQCPlot}
\alias{testRequiredArguments}
\alias{testRequiredFiles}
\alias{testRequiredFilesWarning}
\alias{umap_factor_plot}
\alias{uniquePhosphositesSummariseLongList}
\alias{uniquePhosphositesSummariseWideList}
\alias{update_n}
\alias{updateFilteringProgressLipidomics}
\alias{updateFilteringProgressMetabolomics}
\alias{updateLipidFiltering}
\alias{updateRuvParameters}
\alias{writeInteractiveVolcanoPlotProteomics}
\alias{writeInteractiveVolcanoPlotProteomicsMain}
\alias{writeInteractiveVolcanoPlotProteomicsWidget}
\title{Legacy compatibility API}
\description{
Compatibility aliases retained from the pre-janitor package surface. New
code should prefer the dedicated function and class reference topics where
available.
}
\keyword{internal}
