% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_normalization_methods.R
\name{normaliseBetweenSamples,LipidomicsAssayData-method}
\alias{normaliseBetweenSamples,LipidomicsAssayData-method}
\title{Normalize Between Samples for LipidomicsAssayData}
\usage{
\S4method{normaliseBetweenSamples}{LipidomicsAssayData}(theObject, normalisation_method = NULL)
}
\arguments{
\item{theObject}{Object of class LipidomicsAssayData}

\item{normalisation_method}{Method to use for normalization. Options are
"cyclicloess", "quantile", "scale", "none". If NULL, the value is retrieved
from the object's configuration arguments (looking for "normalisation_method").}
}
\description{
Normalize Between Samples for LipidomicsAssayData
}
