% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_norm_methods.R
\name{normaliseBetweenSamples,ProteinQuantitativeData-method}
\alias{normaliseBetweenSamples,ProteinQuantitativeData-method}
\title{Normalise between Arrays}
\usage{
\S4method{normaliseBetweenSamples}{ProteinQuantitativeData}(theObject, normalisation_method = NULL)
}
\arguments{
\item{theObject}{Object of class ProteinQuantitativeData}

\item{normalisation_method}{Method to use for normalisation. Options are cyclicloess, quantile, scale, none}
}
\description{
Normalise between Arrays
}
