% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_fasta_helpers.R
\name{parseFastaObject}
\alias{parseFastaObject}
\title{Parse FASTA object from seqinr}
\usage{
parseFastaObject(aa_seq)
}
\arguments{
\item{aa_seq}{AAStringSet object, output from running seqinr}
}
\value{
A table containing the protein evidence, isoform number, uniprot accession without isoform number in the uniprot_acc column, gene name
}
\description{
parse_fasta_object: Parse FASTA headers
}
