% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_correlation_methods.R
\name{pearsonCorForSamplePairs,LipidomicsAssayData-method}
\alias{pearsonCorForSamplePairs,LipidomicsAssayData-method}
\title{Calculate Pearson Correlation for Sample Pairs}
\usage{
\S4method{pearsonCorForSamplePairs}{LipidomicsAssayData}(
  theObject,
  tech_rep_remove_regex = NULL,
  correlation_group = NA
)
}
\arguments{
\item{theObject, tech_rep_remove_regex, correlation_group}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Calculate Pearson Correlation for Sample Pairs
}
