% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{plotBarplotMissingnessPerReplicate}
\alias{plotBarplotMissingnessPerReplicate}
\title{plotBarplotMissingnessPerReplicate}
\usage{
plotBarplotMissingnessPerReplicate(
  input_table,
  percent_missing_column = percent_missing,
  fill_column = Machine,
  fill_colour_values = getCmriMachineColour()
)
}
\description{
Plot missing rate for each replicate sample in a bar plot. Colour the bar according to the 'fill_column'.
Depends on the output of the function \code{calculatePercentMissingPerReplicate}
}
