% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_objects.R
\name{plotDensity,ggplot2::ggplot-method}
\alias{plotDensity,ggplot2::ggplot-method}
\title{Create Density Plots from PCA data}
\usage{
\S4method{plotDensity}{ggplot2::ggplot}(theObject, grouping_variable, title = "", font_size = 8)
}
\description{
This function takes a ggplot object (presumably a PCA plot) and creates
density plots for the first two principal components.
}
