% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_plotting_methods.R
\name{plotDensity,LipidomicsAssayData-method}
\alias{plotDensity,LipidomicsAssayData-method}
\title{Plot Density for LipidomicsAssayData}
\usage{
\S4method{plotDensity}{LipidomicsAssayData}(theObject, grouping_variable, title = "", font_size = 8)
}
\arguments{
\item{theObject, grouping_variable, title, font_size}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Plot Density for LipidomicsAssayData
}
