% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_plotting_methods.R
\name{plotDensity,MetaboliteAssayData-method}
\alias{plotDensity,MetaboliteAssayData-method}
\title{Plot Density for MetaboliteAssayData}
\usage{
\S4method{plotDensity}{MetaboliteAssayData}(theObject, grouping_variable, title = "", font_size = 8)
}
\arguments{
\item{theObject, grouping_variable, title, font_size}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Plot Density for MetaboliteAssayData
}
