% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_qc_support_helpers.R
\name{plotNumOfValuesNoLog}
\alias{plotNumOfValuesNoLog}
\title{Plot the number of values in each sample}
\usage{
plotNumOfValuesNoLog(input_table)
}
\arguments{
\item{input_table}{Data matrix with each row as a protein and each column a sample.}
}
\value{
A ggplot2 bar plot showing the number of missing values per column.
}
\description{
Plot the number of values in each sample
}
