% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_volcano_glimma_helpers.R
\name{plotOneVolcano}
\alias{plotOneVolcano}
\title{Draw the volcano plot, used in publication graphs}
\usage{
plotOneVolcano(
  input_data,
  input_title,
  log_q_value_column = lqm,
  log_fc_column = logFC,
  points_type_label = label,
  points_color = colour,
  q_val_thresh = 0.05,
  log2FC_thresh = 1
)
}
\arguments{
\item{input_data}{Input data with the \code{log_q_value_column}, the \code{log_fc_column}, the \code{points_type_label} and the \code{points_color} columns.}

\item{input_title}{Runtime inputs used by this function; see the usage section for accepted values.}

\item{log_q_value_column}{The name of the column representing the log q-value.}

\item{log_fc_column}{The name of the column representing the log fold-change.}

\item{points_type_label}{A column in input table with the type of points based on log fold-change and q-value (e.g. "Not sig., logFC >= 1" = "orange" , "Sig., logFC >= 1" = "purple" , "Sig., logFC < 1" = "blue" , "Not sig." )}

\item{points_color}{A column in input table with the colour of the points corresponding to each type of points (e.g. orange, purple, blue black, )}

\item{q_val_thresh}{A numerical value specifying the q-value threshold for statistically significant proteins.}

\item{log2FC_thresh}{A numerical value specifying the log fold-change threshold to draw a vertical line}
}
\description{
Draw the volcano plot, used in publication graphs
}
