% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_plotting_methods.R
\name{plotPca,LipidomicsAssayData-method}
\alias{plotPca,LipidomicsAssayData-method}
\title{Plot PCA for LipidomicsAssayData}
\usage{
\S4method{plotPca}{LipidomicsAssayData}(
  theObject,
  grouping_variable,
  shape_variable = NULL,
  label_column = NULL,
  title = NULL,
  font_size = 8
)
}
\arguments{
\item{theObject, grouping_variable, shape_variable, label_column, title, font_size}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Plot PCA for LipidomicsAssayData
}
