% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{plotPca,MetaboliteAssayData-method}
\alias{plotPca,MetaboliteAssayData-method}
\title{Plot PCA for MetaboliteAssayData}
\usage{
\S4method{plotPca}{MetaboliteAssayData}(
  theObject,
  grouping_variable,
  shape_variable = NULL,
  label_column,
  title,
  font_size = 8
)
}
\description{
Plot PCA for MetaboliteAssayData
}
\section{Functions}{
\itemize{
\item \code{plotPca(MetaboliteAssayData)}: Method for MetaboliteAssayData

}}
