% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_correlation_methods.R
\name{plotPearson,LipidomicsAssayData-method}
\alias{plotPearson,LipidomicsAssayData-method}
\title{Plot Pearson Correlation}
\usage{
\S4method{plotPearson}{LipidomicsAssayData}(theObject, tech_rep_remove_regex = "pool", correlation_group = NA)
}
\arguments{
\item{theObject, tech_rep_remove_regex, correlation_group}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Plot Pearson Correlation
}
