% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{plotPearson,MetaboliteAssayData-method}
\alias{plotPearson,MetaboliteAssayData-method}
\title{Plot Pearson Correlation Histogram}
\usage{
\S4method{plotPearson}{MetaboliteAssayData}(theObject, tech_rep_remove_regex = "pool", correlation_group = NA)
}
\description{
Plot Pearson Correlation Histogram
}
\section{Functions}{
\itemize{
\item \code{plotPearson(MetaboliteAssayData)}: Method for MetaboliteAssayData

}}
