% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_qc_utility_helpers.R
\name{plotPeptidesProteinsCountsPerSampleHelper}
\alias{plotPeptidesProteinsCountsPerSampleHelper}
\title{plotPeptidesProteinsCountsPerSampleHelper}
\usage{
plotPeptidesProteinsCountsPerSampleHelper(
  input_table,
  intensity_column = Peptide.RawQuantity,
  protein_id_column = Protein.Ids,
  peptide_id_column = Stripped.Sequence,
  sample_id_column = Run,
  peptide_sequence_column = Stripped.Sequence
)
}
\description{
Plot the number of proteins and peptides identified per sample
}
