% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_plotting_methods.R
\name{plotRle,LipidomicsAssayData-method}
\alias{plotRle,LipidomicsAssayData-method}
\title{Plot RLE for LipidomicsAssayData}
\usage{
\S4method{plotRle}{LipidomicsAssayData}(theObject, grouping_variable, yaxis_limit = c(), sample_label = NULL)
}
\arguments{
\item{theObject, grouping_variable, yaxis_limit, sample_label}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Plot RLE for LipidomicsAssayData
}
