% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{plotRle,MetaboliteAssayData-method}
\alias{plotRle,MetaboliteAssayData-method}
\title{Plot RLE for MetaboliteAssayData}
\usage{
\S4method{plotRle}{MetaboliteAssayData}(theObject, grouping_variable, yaxis_limit = c(), sample_label = NULL)
}
\description{
Plot RLE for MetaboliteAssayData
}
\section{Functions}{
\itemize{
\item \code{plotRle(MetaboliteAssayData)}: Method for MetaboliteAssayData

}}
