% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R
\name{plotVolcanoS4,list-method}
\alias{plotVolcanoS4,list-method}
\title{Plot static volcano plot (without gene names)}
\usage{
\S4method{plotVolcanoS4}{list}(
  objectsList,
  de_q_val_thresh = 0.05,
  qvalue_column = "fdr_qvalue",
  log2fc_column = "logFC"
)
}
\description{
Plot static volcano plot (without gene names)
}
