% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_volcano_glimma_helpers.R
\name{printOneVolcanoPlotWithProteinLabel}
\alias{printOneVolcanoPlotWithProteinLabel}
\title{This function creates a volcano plot with protein labels}
\usage{
printOneVolcanoPlotWithProteinLabel(
  input_table,
  uniprot_table,
  protein_id_column = Protein.Ids,
  uniprot_protein_id_column = Entry,
  gene_name_column = gene_name,
  number_of_genes = 100,
  fdr_threshold = 0.05,
  fdr_column = fdr_qvalue,
  log2FC_column = log2FC,
  input_title = "Proteomics",
  include_protein_label = TRUE,
  max.overlaps = 20
)
}
\arguments{
\item{input_table}{The input table to be used for the volcano plot, contains the protein_id_column, fdr_column and log2FC_column}

\item{uniprot_table}{The uniprot table to be used for the volcano plot, contains the uniprot_protein_id_column and gene_name_column}

\item{protein_id_column}{The column name in the input_table that contains the protein ids (tidyverse format)}

\item{uniprot_protein_id_column}{The column name in the uniprot_table that contains the uniprot protein ids (tidyverse format)}

\item{gene_name_column}{The column name in the uniprot_table that contains the gene names (tidyverse format)}

\item{number_of_genes}{Increasing P-value rank for the number of proteins to display on the volcano plot, default is 100`}

\item{fdr_threshold}{The FDR threshold to use for the volcano plot, default is 0.05}

\item{fdr_column}{The column name in the input_table that contains the FDR values (tidyverse format)}

\item{log2FC_column}{The column name in the input_table that contains the log2FC values (tidyverse format)}

\item{input_title}{The title to use for the volcano plot}

\item{max.overlaps}{The maximum number of overlaps to allow for the protein labels using ggrepel (default is 20)}
}
\value{
A ggplot object with the volcano plot and protein labels
}
\description{
This function creates a volcano plot with protein labels
}
