% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_reporting_dispatch_helpers.R
\name{printPValuesDistribution}
\alias{printPValuesDistribution}
\title{Draw the p-values distribution plot.}
\usage{
printPValuesDistribution(
  selected_data,
  p_value_column = raw_pvalue,
  formula_string = "is_ruv_applied ~ comparison"
)
}
\arguments{
\item{selected_data}{A table that is generated by running the function \code{\link{get_significant_data}}.}

\item{formula_string}{The formula string used in the facet_grid command for the ggplot scatter plot.}

\item{log_p_value_column}{The name of the column representing the p-value.}
}
\description{
Draw the p-values distribution plot.
}
