% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_data_helpers.R
\name{proteinIntensityMatrixPivotLonger}
\alias{proteinIntensityMatrixPivotLonger}
\title{Pivot Protein Matrix Long}
\usage{
proteinIntensityMatrixPivotLonger(
  input_matrix,
  sample_id_column,
  protein_id_column,
  quantity_column
)
}
\arguments{
\item{input_matrix, sample_id_column, protein_id_column, quantity_column}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Pivot protein intensity matrix into long format
}
