% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_core.R
\docType{class}
\name{ProteinQuantitativeData-class}
\alias{ProteinQuantitativeData-class}
\alias{ProteinQuantitativeData}
\title{ProteinQuantitativeData S4 Class}
\description{
An S4 class to store and manage protein-level quantitative data
from proteomics experiments along with experimental design metadata.
}
\section{Slots}{

\describe{
\item{\code{protein_quant_table}}{A data.frame containing protein quantification data}

\item{\code{protein_id_column}}{Character string naming the protein ID column}

\item{\code{design_matrix}}{A data.frame containing experimental design}

\item{\code{protein_id_table}}{A data.frame containing protein ID information}

\item{\code{sample_id}}{Character string naming the sample ID column}

\item{\code{group_id}}{Character string naming the group column}

\item{\code{technical_replicate_id}}{Character string naming the replicate column}

\item{\code{args}}{A list of additional arguments}
}}
