% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_support_helpers.R
\name{proteinTechRepCorrelationHelper}
\alias{proteinTechRepCorrelationHelper}
\title{Calculate protein technical replicate correlation}
\usage{
proteinTechRepCorrelationHelper(
  design_matrix_tech_rep,
  data_matrix,
  protein_id_column = "Protein.Ids",
  sample_id_column = "Sample_ID",
  tech_rep_column = "replicates",
  tech_rep_num_column = "tech_rep_num",
  tech_rep_remove_regex = "pool"
)
}
\arguments{
\item{design_matrix_tech_rep, data_matrix, protein_id_column, sample_id_column, tech_rep_column, tech_rep_num_column, tech_rep_remove_regex}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Calculate protein technical replicate correlation
}
