% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_filtering_helpers.R
\name{removeProteinsWithOnlyOneReplicateHelper}
\alias{removeProteinsWithOnlyOneReplicateHelper}
\title{Remove Proteins with Single Replicate}
\usage{
removeProteinsWithOnlyOneReplicateHelper(
  input_table,
  samples_id_tbl,
  input_table_sample_id_column = Run,
  sample_id_tbl_sample_id_column = ms_filename,
  replicate_group_column = general_sample_info,
  protein_id_column = Protein.Ids,
  quantity_column = Protein.Normalised,
  core_utilisation
)
}
\description{
Remove proteins that have been identified in only one replicate across all patients
(e.g. identified no more than one relpicate in any patient)
}
