% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_replicate_helpers.R
\name{removeProteinWithOnlyOneReplicate}
\alias{removeProteinWithOnlyOneReplicate}
\title{Remove Proteins with Single Replicate}
\usage{
removeProteinWithOnlyOneReplicate(
  input_table,
  samples_id_tbl,
  input_table_sample_id_column = Run,
  sample_id_tbl_sample_id_column = ms_filename,
  replicate_group_column = general_sample_info,
  protein_id_column = Protein.Ids,
  core_utilisation
)
}
\description{
Remove proteins that only have data for one technical replicate for all sample.
This can be repurposed for removing proteins that only have one biological replicates for all experimental groups.
}
