% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_filtering_helpers.R
\name{removeRowsWithMissingValuesPercentHelper}
\alias{removeRowsWithMissingValuesPercentHelper}
\title{Remove rows with missing values}
\usage{
removeRowsWithMissingValuesPercentHelper(
  input_table,
  cols,
  design_matrix,
  sample_id,
  row_id,
  grouping_variable,
  groupwise_percentage_cutoff = 1,
  max_groups_percentage_cutoff = 50,
  proteins_intensity_cutoff_percentile = 1,
  temporary_abundance_column = "Abundance"
)
}
\arguments{
\item{input_table}{An input table with a column containing the row ID and the rest of the columns representing abundance values for each sample.}

\item{cols}{A tidyselect command to select the columns. This includes the functions starts_with(), ends_with(), contains(), matches(), and num_range()}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{sample_id}{The name of the column in design_matrix table that has the sample ID.}

\item{row_id}{A unique ID for each row of the 'input_table' variable.}

\item{grouping_variable}{The name of the column in design_matrix table that has the experimental group.}

\item{groupwise_percentage_cutoff}{The maximum percentage of values below threshold allow in each group for a protein .}

\item{max_groups_percentage_cutoff}{The maximum percentage of groups allowed with too many samples with protein abundance values below threshold.}

\item{proteins_intensity_cutoff_percentile}{The percentile of the protein intensity values to be used as the minimum threshold for protein intensity.}

\item{temporary_abundance_column}{The name of a temporary column to keep the abundance value you want to filter upon}
}
\value{
A list, the name of each element is the sample ID and each element is a vector containing the protein accessions (e.g. row_id) with enough number of values.
}
\description{
Remove rows with missing values
}
