% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_qc.R
\name{resolveDuplicateFeatures,MetaboliteAssayData-method}
\alias{resolveDuplicateFeatures,MetaboliteAssayData-method}
\title{Resolve duplicate features in a metabolite assay data object}
\usage{
\S4method{resolveDuplicateFeatures}{MetaboliteAssayData}(theObject, itsd_pattern_columns = NULL)
}
\description{
Resolve duplicate features in a metabolite assay data object
}
\section{Functions}{
\itemize{
\item \code{resolveDuplicateFeatures(MetaboliteAssayData)}: Method for MetaboliteAssayData

}}
