% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_ruv_methods.R
\name{ruvCancor,LipidomicsAssayData-method}
\alias{ruvCancor,LipidomicsAssayData-method}
\title{RUV Canonical Correlation for LipidomicsAssayData}
\usage{
\S4method{ruvCancor}{LipidomicsAssayData}(
  theObject,
  ctrl = NULL,
  num_components_to_impute = NULL,
  ruv_grouping_variable = NULL
)
}
\arguments{
\item{theObject, ctrl, num_components_to_impute, ruv_grouping_variable}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
RUV Canonical Correlation for LipidomicsAssayData
}
