% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{ruvCancor,MetaboliteAssayData-method}
\alias{ruvCancor,MetaboliteAssayData-method}
\title{Plot RUV Canonical Correlation}
\usage{
\S4method{ruvCancor}{MetaboliteAssayData}(
  theObject,
  ctrl = NULL,
  num_components_to_impute = NULL,
  ruv_grouping_variable = NULL
)
}
\description{
Plot RUV Canonical Correlation
}
\section{Functions}{
\itemize{
\item \code{ruvCancor(MetaboliteAssayData)}: Method for MetaboliteAssayData

}}
