% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_da_results_io.R
\name{saveDaProteinList}
\alias{saveDaProteinList}
\title{Save the list of output tables from differential expression analysis of proteins or phosphopeptides into a file and in a specific directory.}
\usage{
saveDaProteinList(
  list_of_da_tables,
  row_id,
  sort_by_column = fdr_qvalue,
  results_dir,
  file_suffix
)
}
\arguments{
\item{list_of_da_tables}{A list, each element is a table of log fold-change and q-values from differential expression analysis of proteins / phosphopeptides. Each element in the list has a name, usually the name of the pairwise comparison.}

\item{row_id}{Add row ID to the output table based on the name (protein or phosphopeptid ID) of each row}

\item{sort_by_column}{Each table in the list_of_da_tables is sorted in ascending order}

\item{results_dir}{The results directory to store the output file}

\item{file_suffix}{The file suffix string to aadd to the name of each comparison from the list_of_da_tables.}
}
\description{
Save the list of output tables from differential expression analysis of proteins or phosphopeptides into a file and in a specific directory.
}
