% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_runtime_helpers.R
\name{saveTimeRecord}
\alias{saveTimeRecord}
\title{Save Time Record}
\usage{
saveTimeRecord(compute_time_record, step_name, toc_output)
}
\arguments{
\item{compute_time_record}{Data frame to store timing records}

\item{step_name}{Name of the step being timed}

\item{toc_output}{Output from tictoc::toc()}
}
\description{
Save timing information for compute steps
}
