% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny_logging.R
\name{setup_shiny_logger}
\alias{setup_shiny_logger}
\title{Setup the logger for the Shiny App}
\usage{
setup_shiny_logger()
}
\description{
This function configures the 'logger' package to use the custom
Shiny appender. It sets the layout to include colors for different log levels.
It should be called once when the Shiny app starts.
}
