% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_filemgmt_report_render_helpers.R
\name{sourceRmdFile}
\alias{sourceRmdFile}
\title{https://gist.github.com/noamross/a549ee50e8a4fd68b8b1
Source the R code from an knitr file, optionally skipping plots}
\usage{
sourceRmdFile(file, skip_plots = TRUE)
}
\arguments{
\item{file}{the knitr file to source}

\item{skip_plots}{whether to make plots. If TRUE (default) sets a null graphics device}
}
\value{
This function is called for its side effects
}
\description{
https://gist.github.com/noamross/a549ee50e8a4fd68b8b1
Source the R code from an knitr file, optionally skipping plots
}
