% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_enrichment_pathway_helpers.R
\name{uniprotGoIdToTermSimple}
\alias{uniprotGoIdToTermSimple}
\title{Convert UniProt GO IDs to terms without grouping or pivoting}
\usage{
uniprotGoIdToTermSimple(
  uniprot_dat,
  uniprot_id_column = UNIPROTKB,
  go_id_column = `GO-IDs`,
  gene_name_column = Gene.Names,
  sep = "; ",
  goterms = AnnotationDbi::Term(GO.db::GOTERM),
  gotypes = AnnotationDbi::Ontology(GO.db::GOTERM)
)
}
\arguments{
\item{uniprot_dat}{a table with uniprot accessions and a column with GO-ID}

\item{uniprot_id_column}{The name of the column with the uniprot accession, as a tidyverse header format, not a string}

\item{go_id_column}{The name of the column with the GO-ID, as a tidyverse header format, not a string}

\item{goterms}{Output from running \code{goterms <- Term(GOTERM)} from the GO.db library.}

\item{gotypes}{Output from running \code{gotypes <- Ontology(GOTERM)} from the GO.db library.}
}
\value{
A table with columns for uniprot_id, go_id, go_term, and go_type
}
\description{
Convert UniProt GO IDs to terms without grouping or pivoting
}
