% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_protein_cleaning_helpers.R
\name{updateProteinIDs}
\alias{updateProteinIDs}
\title{Update Protein IDs based on Sequence Table}
\usage{
updateProteinIDs(protein_data, aa_seq_tbl_final)
}
\arguments{
\item{protein_data}{Data frame containing a 'Protein.Ids' column.}

\item{aa_seq_tbl_final}{Data frame containing accessions and standardized IDs.}
}
\value{
Updated protein_data data frame.
}
\description{
Matches Protein.Ids in the data with accessions in the sequence table
and replaces them with standardized IDs (e.g., database_id or ncbi_refseq).
}
