% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_results}
\alias{write_results}
\title{Write results to a file}
\usage{
write_results(data, filename)
}
\arguments{
\item{data}{The data to be written}

\item{filename}{The name of the file to write the data to}
}
\value{
This function is called for its side effects (writing a file)
}
\description{
This function writes data to a file in the protein_qc subdirectory of the results directory.
}
