% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinVsSamplesS4Objects.R
\name{averageTechReps}
\alias{averageTechReps}
\alias{averageTechReps,ProteinQuantitativeData-method}
\title{Average Technical Replicates}
\usage{
averageTechReps(theObject, design_matrix_columns)

\S4method{averageTechReps}{ProteinQuantitativeData}(theObject, design_matrix_columns = c())
}
\arguments{
\item{theObject}{The object to be processed}

\item{design_matrix_columns}{The columns to be used in the design matrix}

\item{protein_id_column}{The column name of the protein id}

\item{sample_id}{The column name of the sample id}

\item{replicate_group_column}{The column name of the technical replicate id}
}
\description{
Average Technical Replicates
}
