% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R, R/func_prot_qc.R
\name{averageValuesFromReplicates}
\alias{averageValuesFromReplicates}
\title{Average values from replicates}
\usage{
averageValuesFromReplicates(
  input_table,
  design_matrix,
  group_pattern,
  row_id,
  sample_id,
  average_replicates_id
)

averageValuesFromReplicates(
  input_table,
  design_matrix,
  group_pattern,
  row_id,
  sample_id,
  average_replicates_id
)
}
\arguments{
\item{design_matrix}{Contains the sample_id column and the average_replicates_id column}
}
\description{
Average values from replicates

Average values from replicates
}
