% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{avgReplicateProteinIntensity}
\alias{avgReplicateProteinIntensity}
\title{Average Replicate Protein Intensity}
\usage{
avgReplicateProteinIntensity(
  input_table,
  metadata_table,
  protein_id_column = protein_id_column,
  input_table_sample_id_column = Run,
  sample_id_tbl_sample_id_column = Run,
  replicate_group_column = collaborator_patient_id,
  quantity_column = Log2.Protein.Imputed,
  avg_quantity_column = Avg.Log2.Protein.Imputed
)
}
\description{
Protein average values from replicate samples
}
