% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_rollup_count_helpers.R
\name{calcTotalPeptides}
\alias{calcTotalPeptides}
\title{Calculate total unique peptides}
\usage{
calcTotalPeptides(data)
}
\arguments{
\item{data}{A data frame or S4 object containing protein/peptide data}
}
\value{
Integer count of unique peptide-protein combinations
}
\description{
Calculate total unique peptides
}
