% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc.R
\name{calulatePearsonCorrelation}
\alias{calulatePearsonCorrelation}
\title{Calculate Pearson Correlation}
\usage{
calulatePearsonCorrelation(
  ms_filename_x,
  ms_filename_y,
  input_table,
  sample_id_column = "Run",
  protein_id_column = "Protein.Ids",
  peptide_sequence_column = "Stripped.Sequence",
  peptide_normalised_column = "Peptide.Normalised"
)
}
\arguments{
\item{ms_filename_x}{A string representing the sample file name X (for a pair of sample in the same technical replicate group) for correlation score calculation.}

\item{ms_filename_y}{A string representing the sample file name Y (for a pair of sample in the same technical replicate group) for correlation score calculation.}

\item{input_table}{A data frame with the following columns: 1. Sample file name or Run name, 2. Protein IDs, 3. Stripped peptide sequence, 4. Normalised peptide abundances}

\item{sample_id_column}{Sample ID column name as string (default = "Run").}

\item{protein_id_column}{Protein accession column name as string (default = "Protein.Ids").}

\item{peptide_sequence_column}{Peptide sequence column name as string (default = "Stripped.Sequence").}

\item{peptide_normalised_column}{Normalised peptide abundance column name as string (default = "Peptide.Normalised").}
}
\value{
The pearson correlation value of the abundances of peptides between two samples X and Y.
}
\description{
Calculate the Pearson correlation of the abundances of peptides between two samples X and Y.
}
