% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_helpers.R, R/qc_and_rollup.R
\name{changeToCategorical}
\alias{changeToCategorical}
\title{getOneContinousPalette}
\usage{
changeToCategorical(metadata_tbl, column_name, num_colours = 9)

changeToCategorical(metadata_tbl, column_name, num_colours = 9)
}
\description{
getOneContinousPalette

getOneContinousPalette
}
