% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_parameter_helpers.R
\name{checkParamsObjectFunctionSimplify}
\alias{checkParamsObjectFunctionSimplify}
\title{Check the parameters in the arguments list and the function parameters to see what param applies}
\usage{
checkParamsObjectFunctionSimplify(
  theObject,
  param_name_string,
  default_value = NULL
)
}
\arguments{
\item{theObject, param_name_string, default_value}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Check the parameters in the arguments list and the function parameters to see what param applies
}
