% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkParamsObjectFunctionSimplifyAcceptNull}
\alias{checkParamsObjectFunctionSimplifyAcceptNull}
\title{Check the parameters in the arguments list and the function parameters to see what param applies}
\usage{
checkParamsObjectFunctionSimplifyAcceptNull(
  theObject,
  param_name_string,
  default_value = NULL
)
}
\description{
Check the parameters in the arguments list and the function parameters to see what param applies
}
