% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_objects.R
\name{chooseBestProteinAccessionSumDuplicates,ProteinQuantitativeData-method}
\alias{chooseBestProteinAccessionSumDuplicates,ProteinQuantitativeData-method}
\title{Choose Best Protein Accession Sum Duplicates}
\usage{
\S4method{chooseBestProteinAccessionSumDuplicates}{ProteinQuantitativeData}(
  theObject,
  delim = ";",
  quant_columns_pattern = "\\\\d+",
  islogged = TRUE
)
}
\description{
Choose Best Protein Accession Sum Duplicates
}
