% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_objects.R
\name{cleanDesignMatrix,LipidomicsAssayData-method}
\alias{cleanDesignMatrix,LipidomicsAssayData-method}
\title{Clean Design Matrix for LipidomicsAssayData}
\usage{
\S4method{cleanDesignMatrix}{LipidomicsAssayData}(theObject)
}
\description{
Clean Design Matrix for LipidomicsAssayData
}
