% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{cleanDesignMatrix,MetaboliteAssayData-method}
\alias{cleanDesignMatrix,MetaboliteAssayData-method}
\title{Clean Design Matrix for MetaboliteAssayData}
\usage{
\S4method{cleanDesignMatrix}{MetaboliteAssayData}(theObject)
}
\description{
Clean Design Matrix for MetaboliteAssayData
}
