% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation.R
\name{cleanIsoformNumber}
\alias{cleanIsoformNumber}
\title{Clean Isoform Number}
\usage{
cleanIsoformNumber(string)
}
\arguments{
\item{string}{Character vector of protein IDs.}
}
\value{
Cleaned protein IDs with isoform numbers removed.
}
\description{
Remove isoform numbers from protein IDs.
}
