% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_rollup_count_helpers.R
\name{countUniqueProteins}
\alias{countUniqueProteins}
\title{Count unique proteins in peptide or protein data}
\usage{
countUniqueProteins(data)
}
\arguments{
\item{data}{A data frame or S4 object containing protein/peptide data}
}
\value{
Integer count of unique proteins
}
\description{
Count unique proteins in peptide or protein data
}
