% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation.R
\name{createEmptyUniprotTable}
\alias{createEmptyUniprotTable}
\title{Create Empty UniProt Table}
\usage{
createEmptyUniprotTable()
}
\value{
Empty data frame with standard UniProt columns
}
\description{
Creates an empty table with standard UniProt columns when download fails.
}
\keyword{internal}
