% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_enrichment.R
\name{createWordCloudDataFrame}
\alias{createWordCloudDataFrame}
\title{Create a word frequency distribution table for Word Cloud generation.}
\usage{
createWordCloudDataFrame(text_list)
}
\arguments{
\item{text_list, }{a vector of text (e.g. a list of GO terms name)}
}
\description{
Create a word frequency distribution table for Word Cloud generation.
Based on article by Céline Van den Rul, How to Generate Word Clouds in R, Simple Steps on How and When to Use Them,
https://towardsdatascience.com/create-a-word-cloud-with-r-bde3e7422e8a (accessed 7th November 2022)
}
