% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{createWorkflowArgsFromConfig}
\alias{createWorkflowArgsFromConfig}
\title{Create Workflow Arguments Container from Config}
\usage{
createWorkflowArgsFromConfig(
  workflow_name,
  description = "",
  organism_name = NULL,
  taxon_id = NULL,
  source_dir_path = NULL
)
}
\arguments{
\item{workflow_name}{Name of the workflow}

\item{description}{Optional description of the workflow}

\item{organism_name}{Optional organism name (defaults to value from session if available)}

\item{taxon_id}{Optional taxon ID (defaults to value from session if available)}

\item{source_dir_path}{Character string, the path to the source directory for this workflow instance.}
}
\value{
WorkflowArgs object
}
\description{
Create Workflow Arguments Container from Config
}
