% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_da.R
\name{da_server_init_handlers}
\alias{da_server_init_handlers}
\title{Internal initialization handler for Proteomics DA module}
\usage{
da_server_init_handlers(
  input,
  output,
  session,
  da_data,
  workflow_data,
  selected_tab
)
}
\description{
Internal initialization handler for Proteomics DA module
}
\keyword{internal}
