% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_da.R
\name{differentialAbundanceAppletModule}
\alias{differentialAbundanceAppletModule}
\alias{mod_prot_da_ui}
\alias{mod_prot_da_server}
\title{differentialAbundanceAppletModule}
\usage{
mod_prot_da_ui(id)

mod_prot_da_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label,
  selected_tab = NULL
)
}
\description{
A Shiny module for the Differential Abundance step of the proteomics
workflow. Handles limma-based statistical analysis and visualization of results
across user-defined contrasts.
}
