% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{.cleanProteinIds}
\alias{.cleanProteinIds}
\title{Helper function to clean up protein IDs}
\usage{
.cleanProteinIds(ids)
}
\arguments{
\item{ids}{Vector of protein IDs to clean}
}
\value{
Cleaned protein IDs
}
\description{
Helper function to clean up protein IDs
}
