% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_da_model_stats.R
\name{ebFit}
\alias{ebFit}
\title{Run the Empircal Bayes Statistics for Differential Expression in the limma package}
\usage{
ebFit(data, design, contr.matrix)
}
\arguments{
\item{design}{Output from running the function \code{\link{model.matrix}}.}

\item{contr.matrix}{Output from the function \code{\link{makeContrasts}}.}

\item{ID}{List of protein accessions / row names.}
}
\description{
Run the Empircal Bayes Statistics for Differential Expression in the limma package
}
\seealso{
\code{\link{model.matrix}}

\code{\link{makeContrasts}}
}
