% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_enrich.R
\name{enrichmentAnalysisAppletModule}
\alias{enrichmentAnalysisAppletModule}
\alias{mod_prot_enrich_ui}
\alias{mod_prot_enrich_server}
\title{enrichmentAnalysisAppletModule}
\usage{
mod_prot_enrich_ui(id)

mod_prot_enrich_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label,
  selected_tab = NULL
)
}
\description{
A Shiny module for the Enrichment Analysis step of the proteomics
workflow. Handles GO, KEGG, Reactome, and String-DB pathway enrichment analysis
using results from differential expression analysis.
}
