% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideVsSamplesS4Objects.R
\name{filterMinNumPeptidesPerProtein,PeptideQuantitativeData-method}
\alias{filterMinNumPeptidesPerProtein,PeptideQuantitativeData-method}
\title{Filter the proteins based on the number of peptides and peptidoforms}
\usage{
\S4method{filterMinNumPeptidesPerProtein}{PeptideQuantitativeData}(theObject, ...)
}
\arguments{
\item{theObject}{Object of class PeptideQuantitativeData}

\item{num_peptides_per_protein_thresh}{Minimum number of peptides per protein}

\item{num_peptidoforms_per_protein_thresh}{Minimum number of peptidoforms per protein}

\item{core_utilisation}{core_utilisation to use for parallel processing}
}
\description{
Keep the proteins only if they have two or more peptides.
}
