% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_peptide.R
\name{filterMinNumPeptidesPerProteinHelper}
\alias{filterMinNumPeptidesPerProteinHelper}
\title{Filter Proteins by Minimum Number of Peptides}
\usage{
filterMinNumPeptidesPerProteinHelper(
  input_table,
  num_peptides_per_protein_thresh = 1,
  num_peptidoforms_per_protein_thresh = 2,
  protein_id_column = Protein.Ids,
  core_utilisation
)
}
\arguments{
\item{input_table}{Peptide quantities table in long format}

\item{num_peptides_per_protein_thresh}{Minimum number of peptides per protein}

\item{num_peptidoforms_per_protein_thresh}{Minimum number of peptidoforms per protein}

\item{protein_id_column}{Protein ID column name as string}

\item{core_utilisation}{core_utilisation to use for parallel processing}
}
\description{
Keep the proteins only if they have two or more peptides.
}
