% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{filterMinNumPeptidesPerSampleHelper}
\alias{filterMinNumPeptidesPerSampleHelper}
\title{Filter Samples by Minimum Number of Peptides}
\usage{
filterMinNumPeptidesPerSampleHelper(
  input_table,
  peptides_per_sample_cutoff = 5000,
  sample_id_column = Run,
  core_utilisation,
  inclusion_list = c()
)
}
\arguments{
\item{List}{of samples to keep regardless of how many peptides it has because it is am important sample}
}
\description{
Remove sample if it has less than a certain number of peptides identified
}
