% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{filterSamplesByMetaboliteCorrelationThreshold}
\alias{filterSamplesByMetaboliteCorrelationThreshold}
\title{Filter Samples by Metabolite Correlation Threshold}
\usage{
filterSamplesByMetaboliteCorrelationThreshold(
  theObject,
  pearson_correlation_per_pair = NULL,
  min_pearson_correlation_threshold = 0.5
)
}
\description{
Filters samples from a MetaboliteAssayData object based on
pairwise Pearson correlation results. Samples with correlation
below the threshold are removed.
}
