% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc.R, R/qc_and_rollup.R
\name{filterSamplesByProteinCorrelationThresholdHelper}
\alias{filterSamplesByProteinCorrelationThresholdHelper}
\title{Filter Samples by Protein Correlation}
\usage{
filterSamplesByProteinCorrelationThresholdHelper(
  pearson_correlation_per_pair,
  protein_intensity_table,
  min_pearson_correlation_threshold = 0.75,
  filename_column_x = ms_filename.x,
  filename_column_y = ms_filename.y,
  protein_id_column = Protein.Ids,
  correlation_column = pearson_correlation
)

filterSamplesByProteinCorrelationThresholdHelper(
  pearson_correlation_per_pair,
  protein_intensity_table,
  min_pearson_correlation_threshold = 0.75,
  filename_column_x = ms_filename.x,
  filename_column_y = ms_filename.y,
  protein_id_column = Protein.Ids,
  correlation_column = pearson_correlation
)
}
\arguments{
\item{protein_intensity_table}{A data frame with the following columns: 1. ID of technical replicate group, 2. sample file name X, 3. sample file name Y, 4. Pearson correlation of the abundances of peptides between sample X and Y.}

\item{min_pearson_correlation_threshold}{Minimum pearson correlation for a pair of files to be considered to be consistent and kept for further analysis}

\item{filename_column_x}{Name of column containing the sample file name X (for a pair of sample in the same technical replicate group). Tidyverse column header format, not a string.}

\item{filename_column_y}{Name of column containing the sample file name Y (for a pair of sample in the same technical replicate group). Tidyverse column header format, not a string.}

\item{protein_id_column}{Name of column containing the protein ID. Tidyverse column header format, not a string.}

\item{correlation_column}{Name of column containing the Pearson's correlation score between Sample X and Y. Tidyverse column header format, not a string.}

\item{peptide_keep_samples_with_min_num_peptides}{A data frame with the proteins as rows and samples ID as columns.}

\item{filename_id_column}{A string indicating the name of column that contains the sample ID or Run ID in the data frame \code{peptide_keep_samples_with_min_num_peptides}.}
}
\value{
A table without samples that are poorly correlated with the rest of the samples in the technical replicate group. Contains the following columns: 1. Sample file name or Run name, 2. Protein IDs, 3. Stripped peptide sequence, 4. Normalised peptide abundances

A table without samples that are poorly correlated with the rest of the samples in the technical replicate group. Contains the following columns: 1. Sample file name or Run name, 2. Protein IDs, 3. Stripped peptide sequence, 4. Normalised peptide abundances
}
\description{
Remove samples which is correlated with any technical replicate samples

Remove samples which is correlated with any technical replicate samples
}
