% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_import.R
\name{formatDIANNParquet}
\alias{formatDIANNParquet}
\title{Format DIANN Parquet Data}
\usage{
formatDIANNParquet(data_tbl_parquet_filt)
}
\arguments{
\item{data_tbl_parquet_filt}{An EList object containing proteomics data.}
}
\value{
A long-format data frame.
}
\description{
Converts a limpa EList object containing parquet data into a long-format
data frame compatible with MultiScholaR proteomics import.
}
