% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_objects.R, R/func_prot_de.R
\name{generateDEHeatmap}
\alias{generateDEHeatmap}
\title{Generate DE Results Heatmap with Advanced Clustering}
\usage{
generateDEHeatmap(
  de_results_list,
  selected_contrast = NULL,
  top_n_genes = 50,
  clustering_method = "ward.D2",
  distance_method = "euclidean",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  scale_data = "row",
  color_scheme = "RdBu",
  show_gene_names = FALSE,
  de_q_val_thresh = 0.05,
  qvalue_column = "fdr_qvalue",
  log2fc_column = "log2FC"
)

generateDEHeatmap(
  de_results_list,
  selected_contrast = NULL,
  top_n_genes = 50,
  clustering_method = "ward.D2",
  distance_method = "euclidean",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  scale_data = "row",
  color_scheme = "RdBu",
  show_gene_names = FALSE,
  de_q_val_thresh = 0.05,
  qvalue_column = "fdr_qvalue",
  log2fc_column = "log2FC"
)
}
\description{
Generate DE Results Heatmap with Advanced Clustering

Generate DE Results Heatmap with Advanced Clustering
}
