% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_da_volcano_glimma.R
\name{generateProtDAVolcanoPlotGlimma}
\alias{generateProtDAVolcanoPlotGlimma}
\title{Generate Interactive Volcano Plot using Glimma}
\usage{
generateProtDAVolcanoPlotGlimma(
  da_results_list,
  selected_contrast = NULL,
  uniprot_tbl = NULL,
  args_row_id = "uniprot_acc",
  fdr_column = "fdr_qvalue",
  raw_p_value_column = "raw_pvalue",
  log2fc_column = "log2FC",
  da_q_val_thresh = 0.05,
  uniprot_id_column = "Entry",
  gene_names_column = "gene_names",
  display_columns = c("best_uniprot_acc"),
  output_dir = NULL,
  ...
)
}
\description{
Generate Interactive Volcano Plot using Glimma
}
