% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting_palette_theme_helpers.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Generate a color palette}
\usage{
get_color_palette(n, base_color)

get_color_palette(n, base_color)
}
\arguments{
\item{n}{Number of colors needed}

\item{base_color}{Base color to use}
}
\value{
Vector of colors

Vector of colors
}
\description{
Generate a color palette

Generate a color palette
}
