% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{getCategoricalAndContinuousColourRules}
\alias{getCategoricalAndContinuousColourRules}
\title{getCategoricalAndContinuousColourRules}
\usage{
getCategoricalAndContinuousColourRules(
  metadata_tbl,
  metadata_column_labels,
  metadata_column_selected,
  categorical_columns,
  continous_scale_columns,
  ms_machine_column,
  sample_id_column = Run,
  columns_to_exclude,
  na_colour = "white"
)
}
\arguments{
\item{metadata_tbl}{This is the table containing sample ID and other columns containing clinical variables / metadata}

\item{metadata_column_labels}{This is the nice}
}
\description{
getCategoricalAndContinuousColourRules
}
