% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{getCategoricalColourRules}
\alias{getCategoricalColourRules}
\title{getCategoricalColourRules}
\usage{
getCategoricalColourRules(
  metadata_tbl,
  metadata_column_labels,
  metadata_column_selected,
  categorical_columns,
  ms_machine_column,
  columns_to_exclude,
  colour_palette = getCategoricalColourPalette(),
  na_colour = "white"
)
}
\description{
getCategoricalColourRules
}
