% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting.R, R/qc_and_rollup.R
\name{getContinousColourRules}
\alias{getContinousColourRules}
\title{getContinousColourRules}
\usage{
getContinousColourRules(
  metadata_tbl,
  metadata_column_labels,
  metadata_column_selected,
  continous_scale_columns,
  na_colour = "white"
)

getContinousColourRules(
  metadata_tbl,
  metadata_column_labels,
  metadata_column_selected,
  continous_scale_columns,
  na_colour = "white"
)
}
\description{
getContinousColourRules

getContinousColourRules
}
