% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_objects.R
\name{getLowCoefficientOfVariationProteins,ProteinQuantitativeData-method}
\alias{getLowCoefficientOfVariationProteins,ProteinQuantitativeData-method}
\title{Get Low Coefficient of Variation Proteins}
\usage{
\S4method{getLowCoefficientOfVariationProteins}{ProteinQuantitativeData}(
  theObject,
  percentage_as_neg_ctrl = NULL,
  num_neg_ctrl = NULL
)
}
\arguments{
\item{theObject}{A ProteinQuantitativeData object}

\item{percentage_as_neg_ctrl}{Percentage of proteins to use as negative controls (default: 10)}

\item{num_neg_ctrl}{Number of negative control proteins (overrides percentage if set)}
}
\value{
A logical vector indicating which proteins are selected as negative controls
}
\description{
Identifies proteins with the lowest coefficient of variation for use as
negative controls in RUV normalization.
}
