% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{getMaxMinBoxplot}
\alias{getMaxMinBoxplot}
\title{Get Max and Min Boxplot}
\usage{
getMaxMinBoxplot(input_boxplot, adjust_factor = 0.05)
}
\arguments{
\item{input_boxplot}{A ggplot2 boxplot object.}

\item{adjust_factor}{A numeric value to adjust the maximum and minimum data point.}
}
\description{
Input a ggplot2 boxplot, return the maximum and minimum data point adjusted by the adjust_factor.
}
