% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{getNegCtrlProtAnovaHelper}
\alias{getNegCtrlProtAnovaHelper}
\title{Identify negative control proteins for use in removal of unwanted variation, using an ANOVA test.}
\usage{
getNegCtrlProtAnovaHelper(
  data_matrix,
  design_matrix,
  grouping_variable = "group",
  percentage_as_neg_ctrl = 10,
  num_neg_ctrl = round(nrow(data_matrix) * percentage_as_neg_ctrl/100, 0),
  ruv_qval_cutoff = 0.05,
  ruv_fdr_method = "qvalue"
)
}
\arguments{
\item{data_matrix}{A matrix containing the log (base 2) protein abundance values where each column represents a sample and each row represents a protein group, and proteins as rows. The row ID are the protein accessions. The data is preferably median-scaled with missing values imputed.}

\item{design_matrix}{A data frame with the design matrix. Matches sample IDs to group IDs.}

\item{grouping_variable}{The name of the column with the experimental group, as a string.}

\item{num_neg_ctrl}{The number of negative control genes to select. Typically the number of genes with the highest q-value (e.g. least statistically significant). Default is 100}

\item{ruv_qval_cutoff}{The FDR threshold. No proteins with q-values lower than this value are included in the list of negative control proteins. This means the number of negative control proteins could be less than the number specified in \code{num_neg_ctrl} when some were excluded by this threshold.}

\item{ruv_fdr_method}{The FDR calculation method, default is "qvalue". The other option is "BH"}
}
\value{
A boolean vector which indicates which row in the input data matrix is a control gene. The row is included if the value is TRUE. The names of each element is the row ID / protein accessions of the input data matrix.
}
\description{
Identify negative control proteins for use in removal of unwanted variation, using an ANOVA test.
}
