% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{getPairsOfSamplesTable}
\alias{getPairsOfSamplesTable}
\title{Get Pairs of Samples for Comparison}
\usage{
getPairsOfSamplesTable(input_table, run_id_column, replicate_group_column)
}
\arguments{
\item{input_table}{A table with two columns, the Run ID column and the technical replicate group column}

\item{run_id_column}{A string representing the name of the column with the Run ID (or sample ID)}

\item{replicate_group_column}{A string representing the name of the column with the technical replicate group}
}
\value{
A table with three columns, the technical replicate group column, run ID X column, and run ID Y column
}
\description{
Get Pairs of Samples for Comparison
}
