% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_helpers.R
\name{getRHeapMemoryMB}
\alias{getRHeapMemoryMB}
\title{Get R Heap Memory}
\usage{
getRHeapMemoryMB()
}
\value{
Memory usage in MB (numeric)
}
\description{
Returns R's internal heap memory (what pryr::mem_used() shows).
NOTE: This does NOT include environment captures!
}
