% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R, R/func_prot_norm.R
\name{getRuvIIIReplicateMatrixHelper}
\alias{getRuvIIIReplicateMatrixHelper}
\title{Converts a design matrix to a biological replicate matrix for use with ruvIII.}
\usage{
getRuvIIIReplicateMatrixHelper(
  design_matrix,
  sample_id_column,
  grouping_variable,
  temp_column = is_replicate_temp
)

getRuvIIIReplicateMatrixHelper(
  design_matrix,
  sample_id_column,
  grouping_variable,
  temp_column = is_replicate_temp
)
}
\arguments{
\item{design_matrix}{The design matrix with the sample ID in one column and the experimental group in another column}

\item{sample_id_column}{The name of the column with the sample ID, tidyverse style input.}

\item{grouping_variable}{The name of the column with the experimental group, tidyverse style input.}

\item{temp_column}{The name of the temporary column that indicates which samples are biological replicates of the same experimental group.}
}
\value{
A numeric matrix with rows as samples, columns as experimental group, and a value of 1 for samples within the same experimental group represented by the same column, and a value of zero otherwise.

A numeric matrix with rows as samples, columns as experimental group, and a value of 1 for samples within the same experimental group represented by the same column, and a value of zero otherwise.
}
\description{
Converts a design matrix to a biological replicate matrix for use with ruvIII.

Converts a design matrix to a biological replicate matrix for use with ruvIII.
}
