% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/func_general_plotting_heatmap_support_helpers.R
\name{getSamplesCorrelationHeatMap}
\alias{getSamplesCorrelationHeatMap}
\title{getSamplesCorrelationHeatMap}
\usage{
getSamplesCorrelationHeatMap(
  correlation_matrix,
  metadata_tbl,
  is_HEK_column = is_HEK,
  metadata_column_labels,
  metadata_column_selected,
  colour_rules,
  columns_to_exclude,
  sample_id_column = Run,
  use_raster = TRUE,
  raster_device = "CairoPDF",
  heatmap_legend_param = list(title = "Correlation"),
  heatmap_width = ncol(correlation_matrix) * unit(0.05, "cm"),
  heatmap_height = nrow(correlation_matrix) * unit(0.05, "cm")
)
}
\arguments{
\item{correlation_matrix}{Output from the \code{getSamplesCorrelationMatrix} function}

\item{metadata_tbl}{This is the table containing sample ID and other columns containing clinical variables / metadata}

\item{is_HEK_column}{A logical column in the metadata table that indicates if the sample is a HEK sample}

\item{metadata_column_labels}{A list of column names in string to rename each of the columns selected in the param \code{metadata_column_selected}}

\item{metadata_column_selected}{A list of column names in string selected from the metadata tbl}

\item{colour_rules, columns_to_exclude, use_raster, raster_device, heatmap_legend_param, heatmap_width, heatmap_height}{Runtime inputs used by this function; see the usage section for accepted values.}

\item{sample_id_column}{A string describing the column name of the sample ID column}
}
\description{
get the
}
