% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_phospho_annotation.R,
%   R/phosphoproteomics_helper.R
\name{getXMerString}
\alias{getXMerString}
\title{Get X-mer string given the primary sequence and the position of the PTM}
\usage{
getXMerString(seq, uniprot_acc, position, padding_length = 7)

getXMerString(seq, uniprot_acc, position, padding_length = 7)
}
\description{
Given the bioiString object, the uniprot accession, and the position of the phosphorylation site, return the 15mer sequence with the phosphorylation site at the middle.

Given the bioiString object, the uniprot accession, and the position of the phosphorylation site, return the 15mer sequence with the phosphorylation site at the middle.
}
