% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_import.R
\name{importDIANNData}
\alias{importDIANNData}
\title{Import DIA-NN Data}
\usage{
importDIANNData(filepath, use_precursor_norm = TRUE)
}
\arguments{
\item{filepath}{Path to the DIA-NN report file (tsv or parquet).}

\item{use_precursor_norm}{Logical, whether to use Precursor.Normalised instead of Precursor.Quantity.}
}
\value{
A data frame containing the imported DIA-NN data.
}
\description{
Imports a DIA-NN report file, optionally using precursor normalization.
}
