% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_import.R
\name{importDIANNData}
\alias{importDIANNData}
\title{Import DIA-NN Data}
\usage{
importDIANNData(filepath, use_precursor_norm = TRUE)
}
\arguments{
\item{filepath}{Path to DIA-NN report file}

\item{use_precursor_norm}{Whether to use Precursor.Normalised values}
}
\value{
List with data, data_type, and column_mapping
}
\description{
Import DIA-NN Data
}
