% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_pept_norm.R, R/qc_and_rollup.R
\name{log2Transformation}
\alias{log2Transformation}
\title{Log2 Transformation with Pseudo-count}
\usage{
log2Transformation(input_matrix)

log2Transformation(input_matrix)
}
\description{
Log 2 transformation with pseudo count

Log 2 transformation with pseudo count
}
