% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lipid_norm.R
\name{mod_lipid_norm}
\alias{mod_lipid_norm}
\alias{mod_lipid_norm_ui}
\alias{mod_lipid_norm_server}
\title{Lipidomics Normalization Module}
\usage{
mod_lipid_norm_ui(id)

mod_lipid_norm_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label,
  selected_tab = NULL
)
}
\description{
A Shiny module for multi-step normalization of lipidomics data.
Harmonized with proteomics UX pattern.
}
