% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lipid_qc_intensity.R
\name{mod_lipid_qc_intensity}
\alias{mod_lipid_qc_intensity}
\alias{mod_lipid_qc_intensity_ui}
\alias{mod_lipid_qc_intensity_server}
\title{Lipid Intensity Filter Module}
\usage{
mod_lipid_qc_intensity_ui(id)

mod_lipid_qc_intensity_server(id, workflow_data, omic_type, experiment_label)
}
\description{
A Shiny module for applying lipid intensity and missing value filters.
Wraps the lipidIntensityFiltering() S4 method.
}
