% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lipidomics.R
\name{mod_lipidomics}
\alias{mod_lipidomics}
\alias{mod_lipidomics_ui}
\alias{mod_lipidomics_server}
\title{Lipidomics Workflow Orchestrator Module}
\usage{
mod_lipidomics_ui(id)

mod_lipidomics_server(
  id,
  project_dirs,
  omic_type,
  experiment_label,
  volumes = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{project_dirs}{List of project directories from setup}

\item{omic_type}{The omics type (should be "lipidomics")}

\item{experiment_label}{The experiment label for this analysis}

\item{volumes}{Volumes for shinyFiles file browser}
}
\description{
Top-level Shiny module that coordinates the complete lipidomics
workflow including import, design, QC, normalization, differential
analysis, and summary/export.
}
