% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metab_import.R
\name{mod_metab_import}
\alias{mod_metab_import}
\alias{mod_metab_import_ui}
\alias{mod_metab_import_server}
\title{Metabolomics Import Module}
\usage{
mod_metab_import_ui(id)

mod_metab_import_server(id, workflow_data, experiment_paths, volumes = NULL)
}
\description{
A Shiny module for importing metabolomics data with vendor format
detection and dynamic column mapping.
}
