% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_import_server.R
\name{mod_prot_import_server}
\alias{mod_prot_import_server}
\title{Setup & Import Applet Server}
\usage{
mod_prot_import_server(id, workflow_data, experiment_paths, volumes = NULL)
}
\arguments{
\item{id}{Module ID}

\item{workflow_data}{Reactive values object to store workflow data}

\item{experiment_paths}{List of paths for this experiment}

\item{volumes}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Server logic for data import and initial setup with multi-format support
}
