% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_import.R
\name{mod_prot_import_ui}
\alias{mod_prot_import_ui}
\title{Setup & Import Applet UI}
\usage{
mod_prot_import_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
UI for data import and initial setup with multi-format support
}
