% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_impute.R
\name{mod_prot_qc_peptide_impute}
\alias{mod_prot_qc_peptide_impute}
\alias{mod_prot_qc_peptide_impute_ui}
\alias{mod_prot_qc_peptide_impute_server}
\title{Peptide Imputation Module}
\usage{
mod_prot_qc_peptide_impute_ui(id)

mod_prot_qc_peptide_impute_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying missing value imputation.
}
