% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_intensity.R
\name{mod_prot_qc_peptide_intensity}
\alias{mod_prot_qc_peptide_intensity}
\alias{mod_prot_qc_peptide_intensity_ui}
\alias{mod_prot_qc_peptide_intensity_server}
\title{Peptide Intensity Filter Module}
\usage{
mod_prot_qc_peptide_intensity_ui(id)

mod_prot_qc_peptide_intensity_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying peptide intensity filtering.
}
