% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_protein.R
\name{mod_prot_qc_peptide_protein}
\alias{mod_prot_qc_peptide_protein}
\alias{mod_prot_qc_peptide_protein_ui}
\alias{mod_prot_qc_peptide_protein_server}
\title{Protein Peptide Count Filter Module}
\usage{
mod_prot_qc_peptide_protein_ui(id)

mod_prot_qc_peptide_protein_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying the protein peptide count filter.
}
