% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_qvalue.R
\name{mod_prot_qc_peptide_qvalue}
\alias{mod_prot_qc_peptide_qvalue}
\alias{mod_prot_qc_peptide_qvalue_ui}
\alias{mod_prot_qc_peptide_qvalue_server}
\title{Peptide Q-Value Filter Module}
\usage{
mod_prot_qc_peptide_qvalue_ui(id)

mod_prot_qc_peptide_qvalue_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying Q-value and proteotypic peptide filters.
}
