% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_replicate.R
\name{mod_prot_qc_peptide_replicate}
\alias{mod_prot_qc_peptide_replicate}
\alias{mod_prot_qc_peptide_replicate_ui}
\alias{mod_prot_qc_peptide_replicate_server}
\title{Peptide Replicate Filter Module}
\usage{
mod_prot_qc_peptide_replicate_ui(id)

mod_prot_qc_peptide_replicate_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying the replicate filter.
}
