% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_rollup.R
\name{mod_prot_qc_peptide_rollup}
\alias{mod_prot_qc_peptide_rollup}
\alias{mod_prot_qc_peptide_rollup_ui}
\alias{mod_prot_qc_peptide_rollup_server}
\title{Peptide Precursor Rollup Module}
\usage{
mod_prot_qc_peptide_rollup_ui(id)

mod_prot_qc_peptide_rollup_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying precursor to peptide rollup.
}
