% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_peptide_sample.R
\name{mod_prot_qc_peptide_sample}
\alias{mod_prot_qc_peptide_sample}
\alias{mod_prot_qc_peptide_sample_ui}
\alias{mod_prot_qc_peptide_sample_server}
\title{Peptide Sample Filter Module}
\usage{
mod_prot_qc_peptide_sample_ui(id)

mod_prot_qc_peptide_sample_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying the sample quality filter.
}
