% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_cleanup.R
\name{mod_prot_qc_protein_cleanup}
\alias{mod_prot_qc_protein_cleanup}
\alias{mod_prot_qc_protein_cleanup_ui}
\alias{mod_prot_qc_protein_cleanup_server}
\title{Protein Accession Cleanup Module}
\usage{
mod_prot_qc_protein_cleanup_ui(id)

mod_prot_qc_protein_cleanup_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for performing protein accession cleanup.
}
