% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_dedup.R
\name{mod_prot_qc_protein_dedup}
\alias{mod_prot_qc_protein_dedup}
\alias{mod_prot_qc_protein_dedup_ui}
\alias{mod_prot_qc_protein_dedup_server}
\title{Protein Duplicate Removal Module}
\usage{
mod_prot_qc_protein_dedup_ui(id)

mod_prot_qc_protein_dedup_server(
  id,
  workflow_data,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for removing duplicate proteins.
}
