% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_replicate.R
\name{mod_prot_qc_protein_replicate}
\alias{mod_prot_qc_protein_replicate}
\alias{mod_prot_qc_protein_replicate_ui}
\alias{mod_prot_qc_protein_replicate_server}
\title{Protein Replicate Filter Module}
\usage{
mod_prot_qc_protein_replicate_ui(id)

mod_prot_qc_protein_replicate_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for applying the protein replicate filter.
}
