% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_qc_protein_rollup.R
\name{mod_prot_qc_protein_rollup}
\alias{mod_prot_qc_protein_rollup}
\alias{mod_prot_qc_protein_rollup_ui}
\alias{mod_prot_qc_protein_rollup_server}
\title{IQ Protein Rollup Module}
\usage{
mod_prot_qc_protein_rollup_ui(id)

mod_prot_qc_protein_rollup_server(
  id,
  workflow_data,
  experiment_paths,
  omic_type,
  experiment_label
)
}
\description{
A Shiny module for performing peptide-to-protein rollup using the IQ tool.
}
