% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_proteomics.R
\name{mod_proteomics_server}
\alias{mod_proteomics_server}
\title{Proteomics Workflow Server}
\usage{
mod_proteomics_server(
  id,
  project_dirs,
  omic_type,
  experiment_label,
  volumes = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{project_dirs}{Project directory structure from setupDirectories}

\item{omic_type}{The omics type (should be "proteomics")}

\item{experiment_label}{The experiment label for this analysis}

\item{volumes}{Volumes for shinyFiles}
}
\description{
Server logic for the proteomics workflow interface
}
