% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{normaliseBetweenSamples,MetaboliteAssayData-method}
\alias{normaliseBetweenSamples,MetaboliteAssayData-method}
\title{Normalize Between Samples for MetaboliteAssayData}
\usage{
\S4method{normaliseBetweenSamples}{MetaboliteAssayData}(theObject, normalisation_method = NULL)
}
\arguments{
\item{theObject}{Object of class MetaboliteAssayData}

\item{normalisation_method}{Method to use for normalization. Options are
"cyclicloess", "quantile", "scale", "none". If NULL, the value is retrieved
from the object's configuration arguments (looking for "normalisation_method").}
}
\description{
Normalize Between Samples for MetaboliteAssayData
}
