% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideVsSamplesS4Objects.R
\name{normaliseBetweenSamples,PeptideQuantitativeData-method}
\alias{normaliseBetweenSamples,PeptideQuantitativeData-method}
\title{Normalize Between Samples for PeptideQuantitativeData}
\usage{
\S4method{normaliseBetweenSamples}{PeptideQuantitativeData}(theObject, normalisation_method = NULL)
}
\arguments{
\item{theObject}{Object of class PeptideQuantitativeData}

\item{normalisation_method}{Method to use for normalisation. Options are cyclicloess, quantile, scale, none}
}
\description{
Normalize Between Samples for PeptideQuantitativeData
}
