% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R
\name{pearsonCorForSamplePairs,MetaboliteAssayData-method}
\alias{pearsonCorForSamplePairs,MetaboliteAssayData-method}
\title{Calculate Pearson Correlation for Sample Pairs}
\usage{
\S4method{pearsonCorForSamplePairs}{MetaboliteAssayData}(
  theObject,
  tech_rep_remove_regex = NULL,
  correlation_group = NA
)
}
\description{
Calculate Pearson Correlation for Sample Pairs
}
