% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinVsSamplesS4Objects.R
\name{pearsonCorForSamplePairs,ProteinQuantitativeData-method}
\alias{pearsonCorForSamplePairs,ProteinQuantitativeData-method}
\title{Pearson Correlation for Sample Pairs}
\usage{
\S4method{pearsonCorForSamplePairs}{ProteinQuantitativeData}(
  theObject,
  tech_rep_remove_regex = NULL,
  correlation_group = NA
)
}
\arguments{
\item{theObject}{is an object of the type ProteinQuantitativeData}

\item{tech_rep_remove_regex}{samples containing this string are removed from correlation analysis (e.g. if you have lots of pooled sample and want to remove them)}

\item{correlation_group}{is the group where every pair of samples are compared}
}
\description{
Pearson Correlation for Sample Pairs
}
