% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_peptide.R
\name{peptideIntensityFilteringHelper}
\alias{peptideIntensityFilteringHelper}
\title{Filter Peptides by Intensity and Proportion}
\usage{
peptideIntensityFilteringHelper(
  input_table = NULL,
  design_matrix = NULL,
  min_peptide_intensity_threshold = 15,
  sample_id_column = "Run",
  grouping_variable = "Group",
  groupwise_percentage_cutoff = 1,
  max_groups_percentage_cutoff = 50,
  protein_id_column = Protein.Ids,
  peptide_sequence_column = Stripped.Sequence,
  peptide_quantity_column = Peptide.Normalised,
  core_utilisation = NA,
  ...
)
}
\description{
Remove peptide based on the intensity threshold and the proportion of samples below the threshold
}
