% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_peptide.R, R/qc_and_rollup.R
\name{peptideIntensityFilteringHelper}
\alias{peptideIntensityFilteringHelper}
\title{Filter Peptides by Intensity and Proportion}
\usage{
peptideIntensityFilteringHelper(
  input_table,
  min_peptide_intensity_threshold = 15,
  peptides_proportion_of_samples_below_cutoff = 1,
  protein_id_column = Protein.Ids,
  peptide_sequence_column = Stripped.Sequence,
  peptide_quantity_column = Peptide.Normalised,
  core_utilisation
)

peptideIntensityFilteringHelper(
  input_table,
  min_peptide_intensity_threshold = 15,
  peptides_proportion_of_samples_below_cutoff = 1,
  protein_id_column = Protein.Ids,
  peptide_sequence_column = Stripped.Sequence,
  peptide_quantity_column = Peptide.Normalised,
  core_utilisation
)
}
\description{
Remove peptide based on the intensity threshold and the proportion of samples below the threshold

Remove peptide based on the intensity threshold and the proportion of samples below the threshold
}
