% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_peptide_qc_imputation.R
\name{peptideMissingValueImputationHelper}
\alias{peptideMissingValueImputationHelper}
\title{peptideMissingValueImputationHelper}
\usage{
peptideMissingValueImputationHelper(
  input_table,
  metadata_table,
  input_table_sample_id_column = Run,
  sample_id_tbl_sample_id_column = ms_filename,
  replicate_group_column = general_sample_info,
  protein_id_column = Protein.Ids,
  peptide_sequence_column = Stripped.Sequence,
  quantity_to_impute_column = Peptide.Normalised,
  imputed_value_column = Peptide.Imputed,
  hek_string = "HEK",
  proportion_missing_values = 0.5,
  core_utilisation
)
}
\arguments{
\item{input_table}{A data frame with the following columns: 1. Sample file name or Run name, 2. Protein IDs, 3. Stripped peptide sequence, 4. Normalised peptide abundances}

\item{metadata_table}{A data table with the following columns: 1. the sample file name or run name (as per parameter sample_id_tbl_sample_id_column), 2. The replicate group ID (as per parameter replicate_group_column)}

\item{input_table_sample_id_column}{The name of the column in the input_table that contained the run information or sample file name as per the input_table parameter (default: Run)}

\item{sample_id_tbl_sample_id_column}{The name of the column in the input_table that contained the run information or sample file name as per the metadata_table parameter (default: ms_filename)}

\item{replicate_group_column}{(default: general_sample_info)}

\item{protein_id_column}{Protein accession column, tidyverse format (default = Protein.Ids).}

\item{peptide_sequence_column}{Peptide sequence column, tidyverse fromat (default =  Stripped.Sequence).}

\item{quantity_to_impute_column}{Name of column containing the peptide abundance that needs to be normalised in tidyverse format (default: Peptide.RawQuantity)}

\item{hek_string}{The string denoting samples that are controls using HEK cells (default: "HEK")}

\item{proportion_missing_values}{The proportion of sample replicates in a group that is missing below which the peptide intensity will be imputed (default: 0.50)}
}
\description{
Perform peptide level missing value imputation
}
