% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_data_helpers.R
\name{peptidesIntensityMatrixPivotLonger}
\alias{peptidesIntensityMatrixPivotLonger}
\title{Pivot Peptide Matrix Long}
\usage{
peptidesIntensityMatrixPivotLonger(
  input_matrix,
  sample_id_column,
  sequence_column,
  protein_id_column,
  quantity_column,
  unlog_data = TRUE
)
}
\arguments{
\item{input_matrix, sample_id_column, sequence_column, protein_id_column, quantity_column, unlog_data}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Pivot peptide intensity matrix into long format table.
}
