% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_helpers.R
\name{peptidesIntensityMatrixPivotLonger}
\alias{peptidesIntensityMatrixPivotLonger}
\title{Pivot Peptide Matrix Long}
\usage{
peptidesIntensityMatrixPivotLonger(
  input_matrix,
  sample_id_column,
  sequence_column,
  protein_id_column,
  quantity_column,
  unlog_data = TRUE
)
}
\description{
Pivot peptide intensity matrix into long format table.
}
