% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_objects.R
\name{plotDensity,LipidomicsAssayData-method}
\alias{plotDensity,LipidomicsAssayData-method}
\title{Plot Density for LipidomicsAssayData}
\usage{
\S4method{plotDensity}{LipidomicsAssayData}(theObject, grouping_variable, title = "", font_size = 8)
}
\description{
Plot Density for LipidomicsAssayData
}
