% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_s4_plotting_methods.R
\name{plotDensity,list-method}
\alias{plotDensity,list-method}
\title{Plot Density for list of ggplot objects}
\usage{
\S4method{plotDensity}{list}(theObject, grouping_variable, title = "", font_size = 8)
}
\arguments{
\item{theObject, grouping_variable, title, font_size}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Plot Density for list of ggplot objects
}
