% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{plotDensity,list-method}
\alias{plotDensity,list-method}
\title{Plot Density for list of ggplot objects}
\usage{
\S4method{plotDensity}{list}(theObject, grouping_variable, title = "", font_size = 8)
}
\description{
Plot Density for list of ggplot objects
}
