% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R,
%   R/func_lipid_s4_objects.R
\name{plotDensity,list-method}
\alias{plotDensity,list-method}
\title{Plot Density for list of ggplot objects}
\usage{
\S4method{plotDensity}{list}(theObject, grouping_variable, title = "", font_size = 8)

\S4method{plotDensity}{list}(theObject, grouping_variable, title = "", font_size = 8)
}
\description{
Plot Density for list of ggplot objects

Plot Density for list of ggplot objects
}
