% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{plotDensity,MetaboliteAssayData-method}
\alias{plotDensity,MetaboliteAssayData-method}
\title{Plot Density for MetaboliteAssayData}
\usage{
\S4method{plotDensity}{MetaboliteAssayData}(theObject, grouping_variable, title = "", font_size = 8)
}
\description{
Plot Density for MetaboliteAssayData
}
