% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R
\name{plotNumSigDiffExpBarPlot,list-method}
\alias{plotNumSigDiffExpBarPlot,list-method}
\title{plot number of significant differentially expressed metabolites}
\usage{
\S4method{plotNumSigDiffExpBarPlot}{list}(objectsList)
}
\description{
plot number of significant differentially expressed metabolites
}
