% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_objects.R
\name{plotPca,LipidomicsAssayData-method}
\alias{plotPca,LipidomicsAssayData-method}
\title{Plot PCA for LipidomicsAssayData}
\usage{
\S4method{plotPca}{LipidomicsAssayData}(
  theObject,
  grouping_variable,
  shape_variable = NULL,
  label_column = NULL,
  title = NULL,
  font_size = 8
)
}
\description{
Plot PCA for LipidomicsAssayData
}
