% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R
\name{plotPca,MetaboliteAssayData-method}
\alias{plotPca,MetaboliteAssayData-method}
\title{Plot PCA for MetaboliteAssayData}
\usage{
\S4method{plotPca}{MetaboliteAssayData}(
  theObject,
  grouping_variable,
  shape_variable = NULL,
  label_column = NULL,
  title = NULL,
  font_size = 8
)
}
\description{
Plot PCA for MetaboliteAssayData
}
