% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_objects.R
\name{plotPearson,LipidomicsAssayData-method}
\alias{plotPearson,LipidomicsAssayData-method}
\title{Plot Pearson Correlation}
\usage{
\S4method{plotPearson}{LipidomicsAssayData}(theObject, tech_rep_remove_regex = "pool", correlation_group = NA)
}
\description{
Plot Pearson Correlation
}
