% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_s4_objects.R
\name{plotPearson,ProteinQuantitativeData-method}
\alias{plotPearson,ProteinQuantitativeData-method}
\title{Plot Pearson Correlation}
\usage{
\S4method{plotPearson}{ProteinQuantitativeData}(
  theObject,
  tech_rep_remove_regex = NULL,
  correlation_group = NA,
  exclude_pool_samples = TRUE
)
}
\arguments{
\item{theObject}{is an object of the type ProteinQuantitativeData}

\item{tech_rep_remove_regex}{DEPRECATED - use exclude_pool_samples instead}

\item{correlation_group}{is the group where every pair of samples are compared}

\item{exclude_pool_samples}{Logical. If TRUE (default), automatically exclude samples from groups containing "Pool" or "QC" in their name from correlation analysis.}
}
\description{
Plot Pearson Correlation
}
