% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinVsSamplesS4Objects.R
\name{plotPearson,ProteinQuantitativeData-method}
\alias{plotPearson,ProteinQuantitativeData-method}
\title{Plot Pearson Correlation}
\usage{
\S4method{plotPearson}{ProteinQuantitativeData}(theObject, tech_rep_remove_regex = "pool", correlation_group = NA)
}
\arguments{
\item{theObject}{is an object of the type ProteinQuantitativeData}

\item{tech_rep_remove_regex}{samples containing this string are removed from correlation analysis (e.g. if you have lots of pooled sample and want to remove them)}

\item{correlation_group}{is the group where every pair of samples are compared}
}
\description{
Plot Pearson Correlation
}
