% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R
\name{plotRle,MetaboliteAssayData-method}
\alias{plotRle,MetaboliteAssayData-method}
\title{Plot RLE for MetaboliteAssayData}
\usage{
\S4method{plotRle}{MetaboliteAssayData}(theObject, grouping_variable, yaxis_limit = c(), sample_label = NULL)
}
\description{
Plot RLE for MetaboliteAssayData
}
