% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting.R
\name{plotRleHelper}
\alias{plotRleHelper}
\title{Plot RLE}
\usage{
plotRleHelper(
  Y,
  rowinfo = NULL,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  yaxis_limit = c(-0.5, 0.5)
)
}
\arguments{
\item{Y}{Rows = Samples, Columns = Proteins or Peptides}
}
\description{
Plot RLE
}
