% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Draw the volcano plot.}
\usage{
plotVolcano(
  objectsList,
  de_q_val_thresh = 0.05,
  qvalue_column = "q_value",
  log2fc_column = "logFC"
)
}
\arguments{
\item{selected_data}{A table that is generated by running the function \code{\link{get_significant_data}}.}

\item{log_q_value_column}{The name of the column representing the log q-value.}

\item{log_fc_column}{The name of the column representing the log fold-change.}

\item{q_val_thresh}{A numerical value specifying the q-value threshold for statistically significant proteins.}

\item{formula_string}{The formula string used in the facet_grid command for the ggplot scatter plot.}
}
\description{
Draw the volcano plot.
}
