% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R, R/func_prot_de.R
\name{prepareDataForVolcanoPlot}
\alias{prepareDataForVolcanoPlot}
\title{Prepare data for volcano plot}
\usage{
prepareDataForVolcanoPlot(
  input_table,
  protein_id_column = uniprot_acc,
  uniprot_table,
  uniprot_protein_id_column = uniprot_acc_first,
  gene_name_column = gene_name,
  number_of_genes = 3000,
  fdr_threshold = 0.05,
  fdr_column = q.mod,
  log2FC_column = log2FC
)

prepareDataForVolcanoPlot(
  input_table,
  protein_id_column = uniprot_acc,
  uniprot_table,
  uniprot_protein_id_column = uniprot_acc_first,
  gene_name_column = gene_name,
  number_of_genes = 3000,
  fdr_threshold = 0.05,
  fdr_column = q.mod,
  log2FC_column = log2FC
)
}
\arguments{
\item{input_table}{The input table with the log fold-change and q-value columns.}

\item{protein_id_column}{The name of the column representing the protein ID (tidyverse style).}

\item{uniprot_table}{The uniprot table with the imporatnt info on each protein}

\item{uniprot_protein_id_column}{The name of the column representing the protein ID in the uniprot table (tidyverse style).}

\item{number_of_genes}{The number of genes to show in the volcano plot.}

\item{fdr_threshold}{The FDR threshold for the volcano plot.}

\item{fdr_column}{The name of the column representing the FDR value (tidyverse style).}

\item{log2FC_column}{The name of the column representing the log fold-change (tidyverse style).}
}
\value{
A table with the following columns:
label  The label of the significant proteins.
log2FC The log2 fold-change of the significant proteins.
lqm  The -log10 of the q-value.
colour The colour of the significant proteins.
rank_positive: The rank of the positive fold-change values.
rank_negative: The rank of the negative fold-change values.
gene_name_significant  The gene name of the significant proteins.

A table with the following columns:
label  The label of the significant proteins.
log2FC The log2 fold-change of the significant proteins.
lqm  The -log10 of the q-value.
colour The colour of the significant proteins.
rank_positive: The rank of the positive fold-change values.
rank_negative: The rank of the negative fold-change values.
gene_name_significant  The gene name of the significant proteins.
}
\description{
Prepare data for volcano plot

Prepare data for volcano plot
}
