% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/func_general_plotting_reporting_dispatch_helpers.R
\name{printPValuesDistribution}
\alias{printPValuesDistribution}
\title{Draw the p-values distribution plot.}
\usage{
printPValuesDistribution(
  selected_data,
  p_value_column = raw_pvalue,
  formula_string = "is_ruv_applied ~ comparison"
)
}
\arguments{
\item{selected_data}{A table generated by \code{\link[=getSignificantData]{getSignificantData()}}.}

\item{p_value_column}{Runtime inputs used by this function; see the usage section for accepted values.}

\item{formula_string}{The formula string used in the facet_grid command for the ggplot scatter plot.}
}
\description{
Draw the p-values distribution plot.
}
