% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_fasta_processing_helpers.R
\name{processFastaFile}
\alias{processFastaFile}
\title{Process FASTA file and map to UniProt/UniParc}
\usage{
processFastaFile(
  fasta_file_path,
  uniprot_search_results = NULL,
  uniparc_search_results = NULL,
  fasta_meta_file,
  organism_name
)
}
\arguments{
\item{fasta_file_path}{Path to the FASTA file.}

\item{uniprot_search_results}{UniProt search results data frame (optional).}

\item{uniparc_search_results}{UniParc search results data frame (optional).}

\item{fasta_meta_file}{Path to save/load parsed fasta metadata.}

\item{organism_name}{Name of the organism for metadata.}
}
\value{
Mapping table between accessions and standardized IDs.
}
\description{
Parses a FASTA file and maps accessions to UniProt/UniParc search results
to create a standardized sequence mapping table.
}
