% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{proteinIntensityMatrixPivotLonger}
\alias{proteinIntensityMatrixPivotLonger}
\title{Pivot Protein Matrix Long}
\usage{
proteinIntensityMatrixPivotLonger(
  input_matrix,
  sample_id_column,
  protein_id_column,
  quantity_column
)
}
\description{
Pivot protein intensity matrix into long format
}
