% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_support_helpers.R
\name{proteinTechRepCorrelationHelper}
\alias{proteinTechRepCorrelationHelper}
\title{Calculate protein technical replicate correlation}
\usage{
proteinTechRepCorrelationHelper(
  design_matrix_tech_rep,
  data_matrix,
  protein_id_column = "Protein.Ids",
  sample_id_column = "Sample_ID",
  tech_rep_column = "replicates",
  tech_rep_num_column = "tech_rep_num",
  tech_rep_remove_regex = "pool"
)
}
\arguments{
\item{design_matrix_tech_rep:}{design matrix with the technical replicates}

\item{data_matrix:}{input data matrix}

\item{sample_id_column:}{column name of the sample ID. This is the unique identifier for each sample.}

\item{tech_rep_column:}{column name of the technical replicates. Technical replicates of the same sample will have the same value.}

\item{tech_rep_num_column:}{column name of the technical replicate number. This is a unique number for each technical replicate for each sample.}
}
\description{
Calculate protein technical replicate correlation
}
