% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_filemgmt_config_io_helpers.R
\name{readConfigFile}
\alias{readConfigFile}
\title{Read the config file and return the list of parameters}
\usage{
readConfigFile(file = file.path(source_dir, "config.ini"), file_type = "ini")
}
\arguments{
\item{file}{The file path to the config file}

\item{file_type}{The type of the file (default: "ini")}
}
\description{
Read the config file and return the list of parameters
}
