% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_filemgmt_config_io_helpers.R
\name{readConfigFileSection}
\alias{readConfigFileSection}
\title{Read the config file and specify the section and or parameter to update the object}
\usage{
readConfigFileSection(
  theObject,
  file = file.path(source_dir, "config.ini"),
  function_name,
  parameter_name = NULL
)
}
\arguments{
\item{theObject}{The object to be updated}

\item{file}{The file path to the config file}

\item{function_name, parameter_name}{Runtime inputs used by this function; see the usage section for accepted values.}
}
\description{
Read the config file and specify the section and or parameter to update the object
}
