% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{readConfigFileSection}
\alias{readConfigFileSection}
\title{Read the config file and specify the section and or parameter to update the object}
\usage{
readConfigFileSection(
  theObject,
  file = file.path(source_dir, "config.ini"),
  function_name,
  parameter_name = NULL
)
}
\arguments{
\item{theObject}{The object to be updated}

\item{file}{The file path to the config file}

\item{section}{The section to be updated}

\item{value}{The parameter value to be updated}
}
\description{
Read the config file and specify the section and or parameter to update the object
}
