% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_peptide.R, R/qc_and_rollup.R
\name{removePeptidesWithMissingValuesPercentHelper}
\alias{removePeptidesWithMissingValuesPercentHelper}
\title{Remove Peptides with Missing Values}
\usage{
removePeptidesWithMissingValuesPercentHelper(
  input_table,
  design_matrix,
  sample_id,
  protein_id_column,
  peptide_sequence_column,
  grouping_variable,
  groupwise_percentage_cutoff = 1,
  max_groups_percentage_cutoff = 50,
  abundance_threshold,
  abundance_column = "Abundance"
)

removePeptidesWithMissingValuesPercentHelper(
  input_table,
  design_matrix,
  sample_id,
  protein_id_column,
  peptide_sequence_column,
  grouping_variable,
  groupwise_percentage_cutoff = 1,
  max_groups_percentage_cutoff = 50,
  abundance_threshold,
  abundance_column = "Abundance"
)
}
\arguments{
\item{input_table}{An input table with a column containing the row ID and the rest of the columns representing abundance values for each sample.}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{sample_id}{The name of the column in design_matrix table that has the sample ID.}

\item{groupwise_percentage_cutoff}{The maximum percentage of values below threshold allow in each group for a peptide .}

\item{max_groups_percentage_cutoff}{The maximum percentage of groups allowed with too many samples with peptide abundance values below threshold.}

\item{abundance_threshold}{Abundance threshold in which the protein in the sample must be above for it to be considered for inclusion into data analysis.}

\item{cols}{A tidyselect command to select the columns. This includes the functions starts_with(), ends_with(), contains(), matches(), and num_range()}

\item{row_id}{A unique ID for each row of the 'input_table' variable.}

\item{group_column}{The name of the column in design_matrix table that has the experimental group.}

\item{temporary_abundance_column}{The name of a temporary column to keep the abundance value you want to filter upon}
}
\value{
A list, the name of each element is the sample ID and each element is a vector containing the protein accessions (e.g. row_id) with enough number of values.

A list, the name of each element is the sample ID and each element is a vector containing the protein accessions (e.g. row_id) with enough number of values.
}
\description{
Remove Peptides with Missing Values

Remove Peptides with Missing Values
}
