% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_rollup.R
\name{removePeptidesWithOnlyOneReplicateHelper}
\alias{removePeptidesWithOnlyOneReplicateHelper}
\title{Remove Peptides with Single Replicate}
\usage{
removePeptidesWithOnlyOneReplicateHelper(
  input_table,
  samples_id_tbl,
  input_table_sample_id_column = Run,
  sample_id_tbl_sample_id_column = ms_filename,
  replicate_group_column = general_sample_info,
  protein_id_column = Protein.Ids,
  peptide_sequence_column = Stripped.Sequence,
  core_utilisation
)
}
\description{
Remove peptides that only have data for one technical replicate for all sample.
This can be repurposed for removing peptides that only have one biological replicates for all experimental groups.
This function can be repurposed for filtering on proteins as well (we just have to create a dummy variable for peptide_sequence_column)
}
