% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_objects.R
\name{resolveDuplicateFeatures,LipidomicsAssayData-method}
\alias{resolveDuplicateFeatures,LipidomicsAssayData-method}
\title{Resolve Duplicate Features for LipidomicsAssayData}
\usage{
\S4method{resolveDuplicateFeatures}{LipidomicsAssayData}(theObject, itsd_pattern_columns = NULL)
}
\description{
Resolve Duplicate Features for LipidomicsAssayData
}
