% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_metab_s4_objects.R
\name{resolveDuplicateFeatures,MetaboliteAssayData-method}
\alias{resolveDuplicateFeatures,MetaboliteAssayData-method}
\title{Resolve Duplicate Features for MetaboliteAssayData}
\usage{
\S4method{resolveDuplicateFeatures}{MetaboliteAssayData}(theObject, itsd_pattern_columns = NULL)
}
\description{
Resolve Duplicate Features for MetaboliteAssayData
}
