% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_applets.R
\name{RunApplet}
\alias{RunApplet}
\title{Run an applet}
\usage{
RunApplet(
  applet_type,
  omic_type,
  experiment_label,
  project_dirs_object_name = "project_dirs",
  force = FALSE
)
}
\arguments{
\item{applet_type}{The type of applet to run (e.g., "designMatrix").}

\item{omic_type}{The omics context for the applet (e.g., "proteomics", "metabolomics").}

\item{experiment_label}{The label used when setting up directories (e.g., "workshop_data"). This is used to find the correct paths within the project_dirs_object.}

\item{project_dirs_object_name}{The name of the list object in the parent frame that holds the directory structures (typically the output of setupDirectories). Defaults to "project_dirs".}

\item{force}{Logical; if TRUE, skips user confirmation for overwriting existing files (specific to some applets).}
}
\description{
Run an applet
}
