% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_da_model_stats.R
\name{runTests}
\alias{runTests}
\title{Compare a pair of experimental groups and output the log fold-change and q-values per protein.}
\usage{
runTests(
  ID,
  data,
  test_pairs,
  sample_columns,
  sample_rows_list = NA,
  type_of_grouping,
  design_matrix,
  formula_string,
  contrast_variable = "group",
  weights = NA
)
}
\arguments{
\item{ID}{List of protein accessions / row names.}

\item{data}{Data frame containing the log (base 2) protein abundance values where each column represents a sample and each row represents a protein group, and proteins as rows. The data is preferably median-scaled, with missing values imputed, and batch-effects removed.}

\item{test_pairs}{Input file with a table listing all the pairs of experimental groups to compare. First column represents group A and second column represents group B. Linear model comparisons (e.g. Contrasts) would be group B minus group A.}

\item{sample_columns}{A vector of column names (e.g. strings) representing samples which would be used in the statistical tests. Each column contains protein abundance values.}

\item{sample_rows_list}{A list, the name of each element is the sample ID and each element is a vector containing the protein accessions (e.g. row_id) with enough number of values. It is usually the output from \code{\link[=getRowsToKeepList]{getRowsToKeepList()}}.}

\item{type_of_grouping}{A list where each element name is the name of a treatment group and each element is a vector containing the sample IDs within the treatment group. It is usually the output from \code{\link[=getTypeOfGrouping]{getTypeOfGrouping()}}.}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{formula_string}{A formula string representing the experimental design. e.g. ("~ 0 + group")}

\item{contrast_variable}{String representing the contrast variable, which is also used in the formula string. (e.g. "group")}

\item{weights}{Numeric matrix for adjusting each sample and gene.}
}
\value{
A list of data frames, the name of each element represents each pairwise comparison. Each data frame has the following columns:
row.names = the protein accessions
comparison A string showing \verb{log(group B) - log(group A)}.
meanA     mean of the normalised log abundance value of the gene across samples from experimental group A
meanB     mean of the normalised log abundance value of the gene across samples from experimental group B
logFC     log fold-change
tstats    t-test statistics
tmod      moderated t-test statistics
pval      t-test p-value
raw_pvalue      moderated t-test p-value
qval      t-test q-value
fdr_qvalue     moderated t-test q-value
}
\description{
Compare a pair of experimental groups and output the log fold-change and q-values per protein.
}
\seealso{
\code{\link[=getRowsToKeepList]{getRowsToKeepList()}}

\code{\link[=getTypeOfGrouping]{getTypeOfGrouping()}}
}
