% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lipid_s4_objects.R
\name{ruvCancor,LipidomicsAssayData-method}
\alias{ruvCancor,LipidomicsAssayData-method}
\title{RUV Canonical Correlation for LipidomicsAssayData}
\usage{
\S4method{ruvCancor}{LipidomicsAssayData}(
  theObject,
  ctrl = NULL,
  num_components_to_impute = NULL,
  ruv_grouping_variable = NULL
)
}
\description{
RUV Canonical Correlation for LipidomicsAssayData
}
