% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteVsSamplesS4Objects.R
\name{ruvCancor,MetaboliteAssayData-method}
\alias{ruvCancor,MetaboliteAssayData-method}
\title{RUV Canonical Correlation for MetaboliteAssayData}
\usage{
\S4method{ruvCancor}{MetaboliteAssayData}(
  theObject,
  ctrl = NULL,
  num_components_to_impute = NULL,
  ruv_grouping_variable = NULL
)
}
\description{
RUV Canonical Correlation for MetaboliteAssayData
}
