% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_pept_s4_objects.R
\name{ruvCancor,PeptideQuantitativeData-method}
\alias{ruvCancor,PeptideQuantitativeData-method}
\title{RUV Canonical Correlation for PeptideQuantitativeData}
\usage{
\S4method{ruvCancor}{PeptideQuantitativeData}(
  theObject,
  ctrl = NULL,
  num_components_to_impute = NULL,
  ruv_grouping_variable = NULL
)
}
\arguments{
\item{theObject}{A PeptideQuantitativeData object}

\item{ctrl}{Vector of control gene indices for RUV}

\item{num_components_to_impute}{Number of principal components for imputation}

\item{ruv_grouping_variable}{Column name in design matrix for RUV grouping}
}
\description{
RUV Canonical Correlation Analysis for Peptide Data
}
\details{
Performs Remove Unwanted Variation (RUV) using canonical correlation analysis
with DPC imputation for missing values.
}
