% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_plotting.R
\name{save_heatmap_products}
\alias{save_heatmap_products}
\title{Save Heatmap Products}
\usage{
save_heatmap_products(
  heatmap_obj,
  row_clusters,
  params_list,
  output_dir,
  file_prefix
)
}
\arguments{
\item{heatmap_obj}{A ComplexHeatmap object.}

\item{row_clusters}{A named vector of cluster assignments (names are molecule IDs, values are cluster IDs). Can be NULL.}

\item{params_list}{A list of parameters used to generate the heatmap (for documentation).}

\item{output_dir}{The directory where the "Heatmap" subdirectory will be created/used.}

\item{file_prefix}{A string to prefix the filenames with (e.g., "prot_contrast1").}
}
\description{
Saves a heatmap plot to PNG and PDF, writes generation parameters to a Markdown file,
and saves cluster assignments to a CSV file.
}
