% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{savePlot}
\alias{savePlot}
\title{Save a plot in multiple formats}
\usage{
savePlot(
  plot,
  base_path,
  plot_name,
  formats = c("pdf", "png"),
  width = 7,
  height = 7,
  ...
)
}
\arguments{
\item{plot}{The plot object to be saved}

\item{base_path}{The base directory path where the plot will be saved}

\item{plot_name}{The name to be used for the saved plot files}

\item{formats}{A vector of file formats to save the plot in (default: c("pdf", "png"))}

\item{width}{The width of the plot (default: 7)}

\item{height}{The height of the plot (default: 7)}

\item{...}{Additional arguments to be passed to ggsave}
}
\value{
This function is called for its side effects (saving files)
}
\description{
This function saves a given plot in multiple specified formats and also save the ggplot object as a rds file.
}
