% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prot_summary_ui.R and R/mod_prot_summary_server.R
\name{sessionSummaryModule}
\alias{sessionSummaryModule}
\alias{mod_prot_summary_ui}
\alias{mod_prot_summary_server}
\title{sessionSummaryModule}
\usage{
mod_prot_summary_ui(id)

mod_prot_summary_server(
  id,
  project_dirs,
  omic_type = "proteomics",
  experiment_label = NULL,
  workflow_data = NULL
)
}
\description{
A Shiny module for the Session Summary and Report Generation step.
}
