% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_qc_peptide.R
\name{srlQvalueProteotypicPeptideCleanHelper}
\alias{srlQvalueProteotypicPeptideCleanHelper}
\title{Filter Peptides by Q-Value and Proteotypic Status}
\usage{
srlQvalueProteotypicPeptideCleanHelper(
  input_table,
  qvalue_threshold = 0.01,
  global_qvalue_threshold = 0.01,
  choose_only_proteotypic_peptide = 1,
  input_matrix_column_ids = c("Run", "Precursor.Id", "Protein.Ids", "Stripped.Sequence",
    "Modified.Sequence", "Precursor.Charge", "Precursor.Quantity",
    "Precursor.Normalised"),
  protein_id_column = Protein.Ids,
  q_value_column = Q.Value,
  global_q_value_column = Global.Q.Value,
  proteotypic_peptide_sequence_column = Proteotypic
)
}
\description{
Keep spectrum-peptide matches that is within q-value threshold and are proteotypic
}
