% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R, R/func_prot_annotation.R
\name{taxonIdToGprofilerOrganism}
\alias{taxonIdToGprofilerOrganism}
\title{Map NCBI Taxon ID to gprofiler2 Organism Code}
\usage{
taxonIdToGprofilerOrganism(taxon_id)

taxonIdToGprofilerOrganism(taxon_id)
}
\arguments{
\item{taxon_id}{Numeric NCBI taxonomy identifier}
}
\value{
Character string with gprofiler2 organism code

Character string with gprofiler2 organism code
}
\description{
Converts NCBI taxonomy IDs to gprofiler2 organism codes
for use with gprofiler2::gconvert() function

Converts NCBI taxonomy IDs to gprofiler2 organism codes
for use with gprofiler2::gconvert() function
}
\details{
Supported organisms include common model organisms and livestock species.
If the taxon_id is not in the mapping, the function will stop with an error
listing supported organisms.

Supported organisms include common model organisms and livestock species.
If the taxon_id is not in the mapping, the function will stop with an error
listing supported organisms.
}
\examples{
\dontrun{
taxonIdToGprofilerOrganism(9606)  # Returns "hsapiens"
taxonIdToGprofilerOrganism(9940)  # Returns "oaries"
}

\dontrun{
taxonIdToGprofilerOrganism(9606)  # Returns "hsapiens"
taxonIdToGprofilerOrganism(9940)  # Returns "oaries"
}

}
