% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prot_annotation_go_helpers.R
\name{uniprotGoIdToTerm}
\alias{uniprotGoIdToTerm}
\title{Convert UniProt Accession to GO Term}
\usage{
uniprotGoIdToTerm(
  uniprot_dat,
  uniprot_id_column = UNIPROTKB,
  go_id_column = `GO-IDs`,
  sep = "; ",
  goterms = AnnotationDbi::Term(GO.db::GOTERM),
  gotypes = AnnotationDbi::Ontology(GO.db::GOTERM)
)
}
\arguments{
\item{uniprot_dat}{a table with uniprot accessions and a column with GO-ID}

\item{uniprot_id_column}{The name of the column with the uniprot accession, as a tidyverse header format, not a string}

\item{go_id_column}{The name of the column with the GO-ID, as a tidyverse header format, not a string}

\item{sep}{Runtime inputs used by this function; see the usage section for accepted values.}

\item{goterms}{Output from running \code{goterms <- Term(GOTERM)} from the GO.db library.}

\item{gotypes}{Output from running \code{gotypes <- Ontology(GOTERM)} from the GO.db library.}
}
\value{
A table with three columns. go_biological_process, go_celluar_compartment, and go_molecular_function. Each column is a list of gene ontology terms, separated by '; '.
}
\description{
Convert UniProt Accession to GO Term
}
