% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{updateMissingValueParameters}
\alias{updateMissingValueParameters}
\title{Update Missing Value Parameters in Configuration List}
\usage{
updateMissingValueParameters(
  design_matrix,
  config_list,
  min_reps_per_group = 2,
  min_groups = 2
)
}
\arguments{
\item{design_matrix}{A tibble containing the experimental design. Must have a 'group' column.
Groups can have different numbers of replicates.}

\item{config_list}{A list containing configuration parameters. Must have a
'removeRowsWithMissingValuesPercent' nested list with 'groupwise_percentage_cutoff'
and 'max_groups_percentage_cutoff' parameters.}

\item{min_reps_per_group}{Integer specifying the minimum number of replicates required in each passing group.
If a group has fewer total replicates than this value, the minimum is adjusted.}

\item{min_groups}{Integer specifying the minimum number of groups required to have sufficient
quantifiable values. Default is 2.}
}
\value{
Updated config_list with modified missing value parameters
}
\description{
Automatically calculates and updates the missing value filtering parameters in the configuration list
based on the experimental design matrix. The function ensures at least a specified number of groups
have sufficient quantifiable values for analysis.
}
\details{
The function calculates:
\itemize{
\item groupwise_percentage_cutoff: Based on minimum required replicates per group
\item max_groups_percentage_cutoff: Based on minimum required groups
}
}
\examples{
\dontrun{
config_list <- updateMissingValueParameters(design_matrix, config_list, min_reps_per_group = 2)
}

}
