% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_general_helpers.R
\name{updateParamInObject}
\alias{updateParamInObject}
\title{Update the parameter in the object}
\usage{
updateParamInObject(theObject, param_name_string)
}
\description{
Update the parameter in the object
}
