% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_workflow_state.R
\name{WorkflowState}
\alias{WorkflowState}
\title{WorkflowState}
\description{
An R6 class to manage the state of a multi-step workflow.
}
\details{
This class is designed to track the state of data objects (specifically S4 objects)
as they are processed through a series of transformations in a Shiny app. It allows for
saving named states, retrieving them, and reverting to previous states, which is essential
for creating a non-linear, interactive analysis workflow.

It works by storing copies of data objects in a list, acting as a "save game"
feature for the analysis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{states}}{A list to store the different data states. Each element is a
list containing the data object, timestamp, configuration, and other metadata.}

\item{\code{current_state}}{A character string holding the name of the current active state.}

\item{\code{state_history}}{A character vector that logs the sequence of saved states.}

\item{\code{workflow_type}}{A character string indicating the type of workflow (e.g., "LFQ", "TMT").}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkflowState-new}{\code{WorkflowState$new()}}
\item \href{#method-WorkflowState-saveState}{\code{WorkflowState$saveState()}}
\item \href{#method-WorkflowState-getState}{\code{WorkflowState$getState()}}
\item \href{#method-WorkflowState-revertToState}{\code{WorkflowState$revertToState()}}
\item \href{#method-WorkflowState-getHistory}{\code{WorkflowState$getHistory()}}
\item \href{#method-WorkflowState-setWorkflowType}{\code{WorkflowState$setWorkflowType()}}
\item \href{#method-WorkflowState-clone}{\code{WorkflowState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-new}{}}}
\subsection{Method \code{new()}}{
Initialize the WorkflowState object.
Sets up the initial state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-saveState"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-saveState}{}}}
\subsection{Method \code{saveState()}}{
Save a new state (a snapshot of a data object).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$saveState(state_name, s4_data_object, config_object, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state_name}}{A character string for the name of the state (e.g., "raw_data_s4").}

\item{\code{s4_data_object}}{The S4 data object to save. A copy will be made.}

\item{\code{config_object}}{A list containing the configuration parameters used to generate this state.}

\item{\code{description}}{A character string describing the state.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-getState"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-getState}{}}}
\subsection{Method \code{getState()}}{
Retrieve a data object from a specific state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$getState(state_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state_name}}{The name of the state to retrieve. If NULL, retrieves the current state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The data object (e.g., an S4 object) stored in that state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-revertToState"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-revertToState}{}}}
\subsection{Method \code{revertToState()}}{
Revert the current active state to a previously saved state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$revertToState(state_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state_name}}{The name of the state to revert to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The data object from the state that is now active.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-getHistory"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-getHistory}{}}}
\subsection{Method \code{getHistory()}}{
Get the list of all saved state names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$getHistory()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of state names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-setWorkflowType"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-setWorkflowType}{}}}
\subsection{Method \code{setWorkflowType()}}{
Set the workflow type for the session.
Workflow type determines which set of modules to load.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$setWorkflowType(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character string for the workflow type.
Proteomics: "LFQ", "TMT", "DIA"
Metabolomics: "metabolomics_standard"
Lipidomics: "lipidomics_standard"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
